refactor: modularize WordPress plugin abilities
