+++ packages/wordpress-plugin/src/class-wp-codebox-abilities.php
@@ -7,6 +7,16 @@
 
 defined( 'ABSPATH' ) || exit;
 
+require_once __DIR__ . '/trait-wp-codebox-abilities-schemas.php';
+require_once __DIR__ . '/trait-wp-codebox-abilities-execution.php';
+require_once __DIR__ . '/trait-wp-codebox-abilities-permissions.php';
+require_once __DIR__ . '/trait-wp-codebox-abilities-inheritance.php';
+require_once __DIR__ . '/trait-wp-codebox-abilities-browser-artifacts.php';
+require_once __DIR__ . '/trait-wp-codebox-abilities-browser-runtime.php';
+require_once __DIR__ . '/trait-wp-codebox-abilities-browser-blueprint.php';
+require_once __DIR__ . '/trait-wp-codebox-abilities-browser-runner.php';
+require_once __DIR__ . '/trait-wp-codebox-abilities-utils.php';
+
 final class WP_Codebox_Abilities {
 
 	private const BROWSER_ARTIFACT_MAX_BYTES = 5242880;
@@ -15,6 +25,16 @@ final class WP_Codebox_Abilities {
 	private const BROWSER_ARTIFACT_WRITE_SCOPE = 'artifact:write';
 
 	private static bool $registered = false;
+
+	use WP_Codebox_Abilities_Schemas;
+	use WP_Codebox_Abilities_Execution;
+	use WP_Codebox_Abilities_Permissions;
+	use WP_Codebox_Abilities_Inheritance;
+	use WP_Codebox_Abilities_Browser_Artifacts;
+	use WP_Codebox_Abilities_Browser_Runtime;
+	use WP_Codebox_Abilities_Browser_Blueprint;
+	use WP_Codebox_Abilities_Browser_Runner;
+	use WP_Codebox_Abilities_Utils;
 
 	public function __construct() {
 		if ( ! class_exists( 'WP_Ability' ) ) {
@@ -899,3377 +919,5 @@ final class WP_Codebox_Abilities {
 		}
 
 		add_action( 'wp_abilities_api_init', $register_callback );
-	}
-
-	/** @return array<string,mixed> */
-	private static function mount_schema(): array {
-		return array(
-			'type'        => 'array',
-			'description' => 'Additional host directories to mount into the sandbox. Callers may attach repo metadata for downstream tools that need to map sandbox paths back to source repositories.',
-			'items'       => array(
-				'type'       => 'object',
-				'required'   => array( 'source', 'target' ),
-				'properties' => array(
-					'source'   => array(
-						'type'        => 'string',
-						'description' => 'Host directory path to mount.',
-					),
-					'target'   => array(
-						'type'        => 'string',
-						'description' => 'Absolute sandbox path, such as /wordpress/wp-content/plugins/example.',
-					),
-					'mode'     => array(
-						'type' => 'string',
-						'enum' => array( 'readonly', 'readwrite' ),
-					),
-					'metadata' => array(
-						'type'        => 'object',
-						'description' => 'Opaque caller metadata, for example repo, default_branch, repo_root_relative_to_mount, and editable flags.',
-					),
-				),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function site_seed_schema(): array {
-		$scope_schema = array(
-			'type'       => 'object',
-			'properties' => array(
-				'ids'          => array( 'type' => 'array', 'items' => array( 'type' => 'integer' ) ),
-				'slugs'        => array( 'type' => 'array', 'items' => array( 'type' => 'string' ) ),
-				'names'        => array( 'type' => 'array', 'items' => array( 'type' => 'string' ) ),
-				'postTypes'    => array( 'type' => 'array', 'items' => array( 'type' => 'string' ) ),
-				'taxonomies'   => array( 'type' => 'array', 'items' => array( 'type' => 'string' ) ),
-				'roles'        => array( 'type' => 'array', 'items' => array( 'type' => 'string' ) ),
-				'statuses'     => array( 'type' => 'array', 'items' => array( 'type' => 'string' ) ),
-				'includeFiles' => array( 'type' => 'boolean' ),
-				'anonymize'    => array( 'type' => 'boolean' ),
-				'maxRecords'   => array( 'type' => 'integer', 'minimum' => 1, 'maximum' => 100 ),
-			),
-		);
-
-		return array(
-			'type'        => 'array',
-			'description' => 'Explicit opt-in bounded parent-site seed exports for existing-site sandboxes. Exported data is written to a temporary JSON fixture and imported into the sandbox before the task runs.',
-			'items'       => array(
-				'type'       => 'object',
-				'required'   => array( 'type', 'name', 'scopes' ),
-				'properties' => array(
-					'type'   => array( 'type' => 'string', 'enum' => array( 'parent_site' ) ),
-					'name'   => array( 'type' => 'string' ),
-					'scopes' => array(
-						'type'       => 'object',
-						'properties' => array(
-							'posts'         => $scope_schema,
-							'terms'         => $scope_schema,
-							'options'       => $scope_schema,
-							'users'         => $scope_schema,
-							'media'         => $scope_schema,
-							'activePlugins' => array( 'type' => 'boolean' ),
-							'activeTheme'   => array( 'type' => 'boolean' ),
-						),
-					),
-				),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function agent_bundle_schema(): array {
-		return array(
-			'type'        => 'array',
-			'description' => 'Data Machine agent bundles to import into the disposable sandbox before invoking the selected runtime agent.',
-			'items'       => array(
-				'type'       => 'object',
-				'anyOf'      => array(
-					array( 'required' => array( 'source' ) ),
-					array( 'required' => array( 'bundle' ) ),
-				),
-				'properties' => array(
-					'source'      => array(
-						'type'        => 'string',
-						'description' => 'Bundle source accepted by datamachine/import-agent: local directory, .zip, .json, or remote URL.',
-					),
-					'bundle'      => array(
-						'type'        => 'object',
-						'description' => 'Inline Data Machine agent bundle JSON staged in the sandbox and imported through datamachine/import-agent.',
-					),
-					'slug'        => array( 'type' => 'string' ),
-					'on_conflict' => array( 'type' => 'string', 'enum' => array( 'error', 'skip', 'upgrade' ) ),
-					'owner_id'    => array( 'type' => 'integer' ),
-					'token_env'   => array( 'type' => 'string' ),
-				),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function inherit_schema(): array {
-		$credential_secret_schema = array(
-			'type'       => 'object',
-			'required'   => array( 'name', 'status' ),
-			'properties' => array(
-				'name'   => array(
-					'type'        => 'string',
-					'description' => 'Sandbox environment variable name. Secret values are never transported.',
-				),
-				'status' => array(
-					'type' => 'string',
-					'enum' => array( 'available', 'missing', 'denied' ),
-				),
-				'scope'  => array(
-					'type'        => 'string',
-					'description' => 'Parent-side connector scope that authorized this secret name.',
-				),
-				'source' => array(
-					'type'        => 'string',
-					'description' => 'Redacted source label, such as parent-env or connector.',
-				),
-				'reason' => array(
-					'type'        => 'string',
-					'description' => 'Redacted missing/denied reason for audit surfaces.',
-				),
-			),
-		);
-
-		return array(
-			'type'        => 'object',
-			'description' => 'Declarative request for parent-environment connector or setting inheritance. Parent filters resolve names into a sanitized sandbox payload with explicit connector-scoped credential envelopes; secret values are never accepted here.',
-			'properties'  => array(
-				'connectors' => array(
-					'type'        => 'array',
-					'description' => 'Connector names the parent environment should resolve for this sandbox run.',
-					'items'       => array( 'type' => 'string' ),
-				),
-				'settings'   => array(
-					'type'        => 'array',
-					'description' => 'Setting names the parent environment should resolve for artifact/audit metadata. Values are not transported in this slice.',
-					'items'       => array( 'type' => 'string' ),
-				),
-				'credentials' => array(
-					'type'        => 'object',
-					'description' => 'Observable credential envelope shape returned under inheritance.connectors[].credentials by parent filters. Values are not accepted.',
-					'properties'  => array(
-						'schema'    => array( 'type' => 'string' ),
-						'connector' => array( 'type' => 'string' ),
-						'scope'     => array( 'type' => 'string', 'enum' => array( 'connector' ) ),
-						'status'    => array( 'type' => 'string', 'enum' => array( 'available', 'missing', 'denied' ) ),
-						'reason'    => array( 'type' => 'string' ),
-						'secrets'   => array( 'type' => 'array', 'items' => $credential_secret_schema ),
-					),
-				),
-			),
-		);
-	}
-
-	/** @return array<string,array<string,mixed>> */
-	private static function preview_input_schema(): array {
-		return WP_Codebox_Preview_Options::input_schema();
-	}
-
-	/** @return array<string,mixed> */
-	private static function sandbox_session_input_schema(): array {
-		return array(
-			'sandbox_session_id' => array(
-				'type'        => 'string',
-				'description' => 'Caller-owned sandbox session id for external job/session systems. WP Codebox returns it in the response but does not persist sessions itself.',
-			),
-			'orchestrator'       => array(
-				'type'        => 'object',
-				'description' => 'Optional caller-owned job/session metadata, such as external job system, job id, or control-plane id. Values are copied into the response session envelope for correlation only.',
-				'properties'  => array(
-					'id'     => array( 'type' => 'string' ),
-					'type'   => array( 'type' => 'string' ),
-					'job_id' => array( 'type' => 'string' ),
-				),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function remediation_outcome_schema(): array {
-		return array(
-			'type'        => 'object',
-			'description' => 'Strict audit-remediation terminal outcome. Sandbox runs return reviewed artifacts when they can remediate, or an explicit terminal no-op/failure outcome when they cannot. Parent orchestrators apply artifacts and open PRs outside the sandbox.',
-			'properties'  => array(
-				'schema'                => array( 'type' => 'string' ),
-				'success'               => array( 'type' => 'boolean' ),
-				'kind'                  => array(
-					'type' => 'string',
-					'enum' => array( 'fix_artifact', 'false_positive_artifact', 'noop_artifact', 'unable_to_remediate', 'fix_pr', 'false_positive_pr', 'provider_error', 'agent_no_pr_outcome', 'max_turns_exceeded' ),
-				),
-				'failure'               => array(
-					'type' => 'string',
-					'enum' => array( 'provider_error', 'agent_no_pr_outcome', 'max_turns_exceeded', '' ),
-				),
-				'pr_url'                => array( 'type' => 'string' ),
-				'false_positive_pr_url' => array( 'type' => 'string' ),
-				'exit_code'             => array( 'type' => 'integer' ),
-				'retryable'             => array( 'type' => 'boolean' ),
-				'provider_error'        => array( 'type' => 'object' ),
-				'artifact'              => array( 'type' => 'object' ),
-				'metadata'              => array( 'type' => 'object' ),
-				'diagnostics'           => array( 'type' => 'object' ),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function completion_outcome_schema(): array {
-		return array(
-			'type'        => 'object',
-			'description' => 'Generic sandbox completion outcome emitted as files/completion-outcome.json and returned for orchestration without parsing prose.',
-			'properties'  => array(
-				'schema'       => array( 'type' => 'string' ),
-				'status'       => array(
-					'type' => 'string',
-					'enum' => array( 'succeeded', 'blocked', 'failed', 'partial' ),
-				),
-				'summary'      => array( 'type' => 'string' ),
-				'changedFiles' => array( 'type' => 'object' ),
-				'patch'        => array( 'type' => 'object' ),
-				'artifacts'    => array( 'type' => 'object' ),
-				'verification' => array( 'type' => 'object' ),
-				'blockers'     => array( 'type' => 'array' ),
-				'riskNotes'    => array( 'type' => 'array' ),
-				'confidence'   => array( 'type' => 'string' ),
-				'nextAction'   => array( 'type' => 'string' ),
-				'provenance'   => array( 'type' => 'object' ),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function sandbox_session_schema(): array {
-		return array(
-			'type'       => 'object',
-			'properties' => array(
-				'schema'           => array(
-					'type'        => 'string',
-					'description' => 'Session envelope contract version. Use wp-codebox/sandbox-session/v1.',
-				),
-				'id'               => array(
-					'type'        => 'string',
-					'description' => 'Caller-owned sandbox session id echoed for external orchestration correlation.',
-				),
-				'status'           => array(
-					'type'        => 'string',
-					'enum'        => array( 'ready', 'completed' ),
-					'description' => 'Synchronous WP Codebox preparation/run status. Durable queued/running/cancelled/expired state belongs to the external orchestrator.',
-				),
-				'persistence'      => array(
-					'type'        => 'string',
-					'enum'        => array( 'external-orchestrator' ),
-					'description' => 'WP Codebox does not persist host-site session lifecycle state; callers own durable records.',
-				),
-				'agent_session_id' => array( 'type' => 'string' ),
-				'orchestrator'     => array( 'type' => 'object' ),
-				'authorization'    => array( 'type' => 'object' ),
-				'artifacts'        => array( 'type' => 'object' ),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function browser_playground_session_schema(): array {
-		return array(
-			'type'       => 'object',
-			'properties' => array(
-				'success'    => array( 'type' => 'boolean' ),
-				'schema'     => array( 'type' => 'string' ),
-				'execution'  => array(
-					'type'        => 'string',
-					'enum'        => array( 'browser-playground' ),
-					'description' => 'The caller browser executes WordPress Playground; the host site does not run the WP Codebox CLI.',
-				),
-				'execution_scope'  => array(
-					'type'        => 'string',
-					'enum'        => array( 'disposable-playground' ),
-					'description' => 'Browser sessions run inside a disposable WordPress Playground sandbox, not the host site.',
-				),
-				'permission_model' => array(
-					'type'        => 'string',
-					'enum'        => array( 'sandbox-bypass' ),
-					'description' => 'The generated browser runner bypasses agents/chat permission checks only inside the disposable Playground sandbox.',
-				),
-				'session'    => array( 'type' => 'object' ),
-				'task_input' => self::task_input_schema(),
-				'task_payload' => array( 'type' => 'object' ),
-				'playground' => array( 'type' => 'object' ),
-				'runtime'    => array( 'type' => 'object' ),
-				'site_blueprint_artifact' => array( 'type' => 'object' ),
-				'materialization' => array(
-					'type'        => 'object',
-					'description' => 'Generic browser materialization invocation contract and result capture shape produced by the generated Playground runner.',
-				),
-				'recipe'     => array( 'type' => 'object' ),
-				'signals'    => array( 'type' => 'object' ),
-				'artifacts'  => array( 'type' => 'object' ),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function browser_materializer_contract_schema(): array {
-		return array(
-			'type'       => 'object',
-			'properties' => array(
-				'success'          => array( 'type' => 'boolean' ),
-				'schema'           => array( 'type' => 'string' ),
-				'execution'        => array(
-					'type' => 'string',
-					'enum' => array( 'browser-playground' ),
-				),
-				'execution_scope'  => array(
-					'type' => 'string',
-					'enum' => array( 'disposable-playground' ),
-				),
-				'permission_model' => array(
-					'type' => 'string',
-					'enum' => array( 'sandbox-bypass' ),
-				),
-				'session_id'       => array( 'type' => 'string' ),
-				'authorization'    => array( 'type' => 'object' ),
-				'task_input'       => self::task_input_schema(),
-				'task_payload'     => array( 'type' => 'object' ),
-				'materialization'  => array( 'type' => 'object' ),
-				'recipe'           => array( 'type' => 'object' ),
-				'playground'       => array( 'type' => 'object' ),
-				'runtime'          => array( 'type' => 'object' ),
-				'artifacts'        => array( 'type' => 'object' ),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function browser_session_authorization_schema(): array {
-		return self::trusted_orchestrator_authorization_schema( self::BROWSER_SESSION_CREATE_SCOPE, 'Explicit trusted orchestrator authorization for browser session creation. Callers must provide a caller id and the browser-session:create scope; sites grant trust through wp_codebox_trusted_browser_session_callers.' );
-	}
-
-	/** @return array<string,mixed> */
-	private static function trusted_orchestrator_authorization_schema( string $scope, string $description ): array {
-		return array(
-			'type'        => 'object',
-			'description' => $description,
-			'properties'  => array(
-				'schema' => array(
-					'type'        => 'string',
-					'description' => 'Authorization contract version. Use wp-codebox/trusted-orchestrator-authorization/v1.',
-				),
-				'caller' => array(
-					'type'        => 'string',
-					'description' => 'Stable caller id, for example studio-web.',
-				),
-				'scope'  => array(
-					'type'        => 'string',
-					'enum'        => array( $scope ),
-					'description' => 'Required trusted-orchestrator capability scope.',
-				),
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function task_input_schema(): array {
-		return WP_Codebox_Task_Input_Contract::schema();
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function run_agent_task( array $input ): array|WP_Error {
-		return ( new WP_Codebox_Agent_Sandbox_Runner() )->run( $input );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function run_agent_task_batch( array $input ): array|WP_Error {
-		return ( new WP_Codebox_Agent_Sandbox_Runner() )->run_batch( $input );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function create_browser_playground_session( array $input ): array|WP_Error {
-		$task_input = self::normalize_task_input( $input );
-		if ( is_wp_error( $task_input ) ) {
-			return $task_input;
-		}
-
-		$session_id = trim( (string) ( $input['sandbox_session_id'] ?? $input['session_id'] ?? '' ) );
-		if ( '' === $session_id ) {
-			$session_id = self::generate_id();
-		}
-
-		$playground = self::browser_playground( $input );
-		if ( is_wp_error( $playground ) ) {
-			return $playground;
-		}
-
-		$inheritance_payload = self::browser_inheritance_resolution_payload( $input );
-		if ( is_wp_error( $inheritance_payload ) ) {
-			return $inheritance_payload;
-		}
-		$input           = self::browser_input_with_inheritance( $input, $inheritance_payload['inheritance'] );
-		$browser_runner  = is_array( $input['browser_runner'] ?? null ) ? $input['browser_runner'] : array();
-		$legacy_plugins  = self::browser_plugins( $input );
-		if ( is_wp_error( $legacy_plugins ) ) {
-			return $legacy_plugins;
-		}
-		$runtime = self::browser_runtime_dependencies( $input, $legacy_plugins );
-		if ( is_wp_error( $runtime ) ) {
-			return $runtime;
-		}
-		$browser_plugins = $runtime['plugins'];
-
-		$site_blueprint_artifact = self::browser_site_blueprint_artifact( $input );
-		if ( is_wp_error( $site_blueprint_artifact ) ) {
-			return $site_blueprint_artifact;
-		}
-
-		$base_blueprint = self::browser_blueprint_with_site_artifact( is_array( $input['blueprint'] ?? null ) ? $input['blueprint'] : array(), $site_blueprint_artifact );
-		$blueprint      = self::browser_blueprint_with_runtime( $base_blueprint, $runtime, $playground );
-		$artifacts      = self::browser_artifact_files( $input );
-		if ( is_wp_error( $artifacts ) ) {
-			return $artifacts;
-		}
-		$ready_to_code = self::browser_ready_to_code_signal( $input, $runtime );
-		if ( false === ( $ready_to_code['emitted'] ?? false ) ) {
-			return self::blocked_browser_playground_session( $session_id, $input, $task_input, $ready_to_code, $browser_plugins, $runtime, $artifacts, $playground, $blueprint, $site_blueprint_artifact );
-		}
-
-		$task_payload = self::browser_task_payload( $input, $task_input, $session_id, $artifacts, $inheritance_payload['inheritance'] );
-		$recipe = self::browser_agent_recipe( $task_input, $session_id, $browser_runner, $blueprint, $playground, $task_payload );
-		if ( is_wp_error( $recipe ) ) {
-			return $recipe;
-		}
-		$materialization = self::browser_materialization_contract( $recipe );
-
-		return array(
-			'success'          => true,
-			'schema'           => 'wp-codebox/browser-playground-session/v1',
-			'execution'        => 'browser-playground',
-			'execution_scope'  => 'disposable-playground',
-			'permission_model' => 'sandbox-bypass',
-			'session'          => self::browser_session_envelope( $session_id, 'ready', $input ),
-			'task'             => (string) $task_input['goal'],
-			'task_input' => $task_input,
-			'task_payload' => $task_payload,
-			'agent'      => (string) ( $input['agent'] ?? 'wp-codebox-sandbox' ),
-			'provider'   => self::browser_provider( $input, $inheritance_payload['inheritance'] ),
-			'model'      => self::browser_model( $input, $inheritance_payload['inheritance'] ),
-			'inheritance' => $inheritance_payload['inheritance'],
-			'plugins'    => $browser_plugins,
-			'runtime'    => $runtime,
-			'site_blueprint_artifact' => $site_blueprint_artifact,
-			'materialization' => $materialization,
-			'playground' => array(
-				'client_module_url'  => $playground['client_module_url'],
-				'remote_url'         => $playground['remote_url'],
-				'cors_proxy_url'     => $playground['cors_proxy_url'],
-				'scope'              => (string) ( $playground['scope'] ?? $session_id ),
-				'artifact_base_path' => self::browser_artifact_base_path( $playground ),
-				'artifact_base_url'  => self::browser_artifact_base_url( $playground ),
-				'preview_url'        => self::browser_preview_url( $artifacts, $playground ),
-				'blueprint'          => self::browser_playground_blueprint( $blueprint, $playground ),
-				'capabilities'       => array(
-					'compile_blueprint' => true,
-					'run_blueprint'     => true,
-					'write_file'        => true,
-					'run_php'           => true,
-				),
-				'provenance'         => $playground['provenance'],
-			),
-			'recipe'     => $recipe,
-			'signals'    => array(
-				'ready_to_code' => $ready_to_code,
-			),
-			'artifacts'  => array(
-				'schema'             => 'wp-codebox/browser-artifacts/v1',
-				'files'              => $artifacts,
-				'preview_url'        => self::browser_preview_url( $artifacts, $playground ),
-				'expected_artifacts' => $task_input['expected_artifacts'],
-			),
-		);
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function create_browser_materializer_contract( array $input ): array|WP_Error {
-		$session = self::create_browser_playground_session( $input );
-		if ( is_wp_error( $session ) ) {
-			return $session;
-		}
-
-		if ( true !== ( $session['success'] ?? false ) ) {
-			$session_envelope = is_array( $session['session'] ?? null ) ? $session['session'] : array();
-			return array_filter(
-				array(
-					'success'          => false,
-					'schema'           => 'wp-codebox/browser-materializer-contract/v1',
-					'execution'        => 'browser-playground',
-					'execution_scope'  => 'disposable-playground',
-					'permission_model' => 'sandbox-bypass',
-					'status'           => (string) ( $session['status'] ?? 'blocked' ),
-					'error'            => is_array( $session['error'] ?? null ) ? $session['error'] : array(),
-					'session_id'       => (string) ( $session_envelope['id'] ?? '' ),
-					'authorization'    => is_array( $session_envelope['authorization'] ?? null ) ? $session_envelope['authorization'] : self::browser_session_authorization( $input ),
-					'signals'          => is_array( $session['signals'] ?? null ) ? $session['signals'] : array(),
-				),
-				static fn( mixed $value ): bool => array() !== $value && '' !== $value
-			);
-		}
-
-		$session_envelope = is_array( $session['session'] ?? null ) ? $session['session'] : array();
-
-		return array(
-			'success'          => true,
-			'schema'           => 'wp-codebox/browser-materializer-contract/v1',
-			'execution'        => 'browser-playground',
-			'execution_scope'  => 'disposable-playground',
-			'permission_model' => 'sandbox-bypass',
-			'session_id'       => (string) ( $session_envelope['id'] ?? '' ),
-			'authorization'    => is_array( $session_envelope['authorization'] ?? null ) ? $session_envelope['authorization'] : self::browser_session_authorization( $input ),
-			'task_input'       => is_array( $session['task_input'] ?? null ) ? $session['task_input'] : array(),
-			'task_payload'     => is_array( $session['task_payload'] ?? null ) ? $session['task_payload'] : array(),
-			'materialization'  => is_array( $session['materialization'] ?? null ) ? $session['materialization'] : array(),
-			'recipe'           => is_array( $session['recipe'] ?? null ) ? $session['recipe'] : array(),
-			'playground'       => is_array( $session['playground'] ?? null ) ? $session['playground'] : array(),
-			'runtime'          => is_array( $session['runtime'] ?? null ) ? $session['runtime'] : array(),
-			'artifacts'        => is_array( $session['artifacts'] ?? null ) ? $session['artifacts'] : array(),
-			'provenance'       => array(
-				'generated_by' => 'wp-codebox/browser-materializer-contract',
-				'source'       => 'wp-codebox/create-browser-playground-session',
-			),
-		);
-	}
-
-	/**
-	 * @param array<string,mixed> $input Ability input.
-	 * @param array<string,mixed> $task_input Normalized task input.
-	 * @param array<string,mixed> $ready_to_code Readiness signal.
-	 * @param array<int,array<string,string>> $browser_plugins Browser plugin specs.
-	 * @param array<string,mixed> $runtime Normalized runtime dependency specs.
-	 * @param array<int,array<string,string>> $artifacts Browser artifact specs.
-	 * @param array<string,mixed> $playground Playground input.
-	 * @param array<string,mixed> $blueprint Playground blueprint.
-	 * @param array<string,mixed> $site_blueprint_artifact Normalized site blueprint artifact.
-	 * @return array<string,mixed>
-	 */
-	private static function blocked_browser_playground_session( string $session_id, array $input, array $task_input, array $ready_to_code, array $browser_plugins, array $runtime, array $artifacts, array $playground, array $blueprint, array $site_blueprint_artifact ): array {
-		return array(
-			'success'          => false,
-			'schema'           => 'wp-codebox/browser-playground-session/v1',
-			'execution'        => 'browser-playground',
-			'execution_scope'  => 'disposable-playground',
-			'permission_model' => 'sandbox-bypass',
-			'status'           => 'blocked',
-			'error'            => array(
-				'code'    => 'wp_codebox_browser_prerequisites_missing',
-				'message' => 'Browser Playground sandbox is missing required coding prerequisites.',
-				'missing' => $ready_to_code['missing'] ?? array(),
-			),
-			'session'          => self::browser_session_envelope( $session_id, 'blocked', $input ),
-			'task'             => (string) $task_input['goal'],
-			'task_input' => $task_input,
-			'agent'      => (string) ( $input['agent'] ?? 'wp-codebox-sandbox' ),
-			'plugins'    => $browser_plugins,
-			'runtime'    => $runtime,
-			'site_blueprint_artifact' => $site_blueprint_artifact,
-			'materialization' => array(
-				'schema' => 'wp-codebox/browser-materialization/v1',
-				'status' => 'blocked',
-				'captures' => array(),
-			),
-			'playground' => array(
-				'client_module_url'  => $playground['client_module_url'],
-				'remote_url'         => $playground['remote_url'],
-				'cors_proxy_url'     => $playground['cors_proxy_url'],
-				'scope'              => (string) ( $playground['scope'] ?? $session_id ),
-				'artifact_base_path' => self::browser_artifact_base_path( $playground ),
-				'artifact_base_url'  => self::browser_artifact_base_url( $playground ),
-				'preview_url'        => self::browser_preview_url( $artifacts, $playground ),
-				'blueprint'          => self::browser_playground_blueprint( $blueprint, $playground ),
-				'capabilities'       => array(
-					'compile_blueprint' => true,
-					'run_blueprint'     => true,
-					'write_file'        => true,
-					'run_php'           => true,
-				),
-				'provenance'         => $playground['provenance'],
-			),
-			'signals'    => array(
-				'ready_to_code' => $ready_to_code,
-			),
-			'artifacts'  => array(
-				'schema'             => 'wp-codebox/browser-artifacts/v1',
-				'files'              => $artifacts,
-				'preview_url'        => self::browser_preview_url( $artifacts, $playground ),
-				'expected_artifacts' => $task_input['expected_artifacts'],
-			),
-		);
-	}
-
-	/** @return array<string,mixed> */
-	private static function browser_session_envelope( string $session_id, string $status, array $input ): array {
-		$session = WP_Codebox_Agent_Task::session( $session_id, $status, $input );
-		$session['execution_scope']  = 'disposable-playground';
-		$session['permission_model'] = 'sandbox-bypass';
-		$authorization               = self::browser_session_authorization( $input );
-		if ( ! empty( $authorization['caller'] ) || ! empty( $authorization['scope'] ) ) {
-			$session['authorization'] = $authorization;
-		}
-
-		return $session;
-	}
-
-	/** @param array<string,mixed> $input Ability input. @param array<string,mixed> $runtime Normalized runtime dependency specs. @return array<string,mixed> */
-	private static function browser_ready_to_code_signal( array $input, array $runtime ): array {
-		$provider_plugin_paths = array_values(
-			array_filter(
-				array_map( 'strval', is_array( $input['provider_plugin_paths'] ?? null ) ? $input['provider_plugin_paths'] : array() ),
-				static fn( string $path ): bool => '' !== trim( $path )
-			)
-		);
-		$inherit      = is_array( $input['inherit'] ?? null ) ? $input['inherit'] : array();
-		$connectors   = array_values( array_filter( array_map( 'strval', is_array( $inherit['connectors'] ?? null ) ? $inherit['connectors'] : array() ) ) );
-		$secret_env   = array_values( array_filter( array_map( 'strval', is_array( $input['secret_env'] ?? null ) ? $input['secret_env'] : array() ) ) );
-		$requirements = array(
-			'agents_api'        => self::agents_api_ready() && self::browser_runtime_has_plugin( $runtime, 'agents-api' ),
-			'data_machine'      => self::browser_runtime_has_plugin( $runtime, 'data-machine' ),
-			'data_machine_code' => self::browser_runtime_has_plugin( $runtime, 'data-machine-code' ),
-			'provider_plugin'   => ! empty( $provider_plugin_paths ) && self::all_paths_ready( $provider_plugin_paths ),
-			'provider_secret'   => ! empty( $connectors ) || ! empty( $secret_env ),
-			'runtime_dependencies' => true,
-		);
-
-		/**
-		 * Filters browser sandbox readiness requirements before the signal is emitted.
-		 *
-		 * @param array<string,bool>  $requirements Named readiness checks.
-		 * @param array<string,mixed> $input        Ability input.
-		 */
-		$requirements = apply_filters( 'wp_codebox_browser_ready_to_code_requirements', $requirements, $input );
-		$requirements = is_array( $requirements ) ? array_map( 'boolval', $requirements ) : array();
-		$missing      = array_keys( array_filter( $requirements, static fn( bool $ready ): bool => ! $ready ) );
-		$emitted      = empty( $missing );
-
-		return array(
-			'schema'       => 'wp-codebox/signal/v1',
-			'name'         => 'ready_to_code',
-			'emitted'      => $emitted,
-			'message'      => $emitted ? 'Browser Playground sandbox is ready to code.' : 'Browser Playground sandbox is not ready to code.',
-			'requirements' => $requirements,
-			'requirement_metadata' => array(
-				'runtime_dependencies' => self::browser_runtime_readiness_metadata( $runtime ),
-			),
-			'missing'      => $missing,
-		);
-	}
-
-	private static function agents_api_ready(): bool {
-		if ( ! function_exists( 'wp_get_ability' ) ) {
-			return false;
-		}
-
-		return (bool) wp_get_ability( 'agents/chat' );
-	}
-
-	/** @param array<int,string> $paths Paths to verify. */
-	private static function all_paths_ready( array $paths ): bool {
-		foreach ( $paths as $path ) {
-			if ( ! is_dir( $path ) ) {
-				return false;
-			}
-		}
-
-		return true;
-	}
-
-	/** @param array<string,mixed> $runtime Normalized runtime dependencies. */
-	private static function browser_runtime_has_plugin( array $runtime, string $slug ): bool {
-		foreach ( is_array( $runtime['plugins'] ?? null ) ? $runtime['plugins'] : array() as $plugin ) {
-			if ( is_array( $plugin ) && $slug === self::safe_key( (string) ( $plugin['slug'] ?? '' ) ) ) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function list_artifacts( array $input = array() ): array|WP_Error {
-		return ( new WP_Codebox_Artifacts() )->list( $input );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function get_artifact( array $input ): array|WP_Error {
-		return ( new WP_Codebox_Artifacts() )->get( $input );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function discard_artifact( array $input ): array|WP_Error {
-		return ( new WP_Codebox_Artifacts() )->discard( $input );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function normalize_browser_artifact_bundle( array $input ): array|WP_Error {
-		return ( new WP_Codebox_Artifacts() )->normalize_browser_bundle( $input );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function persist_browser_artifact( array $input ): array|WP_Error {
-		$result = ( new WP_Codebox_Artifacts() )->persist_browser_bundle( $input );
-		if ( is_wp_error( $result ) ) {
-			return $result;
-		}
-
-		$authorization = self::trusted_orchestrator_authorization( $input, self::BROWSER_ARTIFACT_WRITE_SCOPE );
-		if ( ! empty( $authorization['caller'] ) || ! empty( $authorization['scope'] ) ) {
-			$result['authorization'] = $authorization;
-		}
-
-		return $result;
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function review_artifact( array $input ): array|WP_Error {
-		return ( new WP_Codebox_Artifacts() )->review_artifact( $input );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function apply_artifact_preflight( array $input ): array|WP_Error {
-		return ( new WP_Codebox_Artifacts() )->apply_preflight( $input );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function apply_approved_artifact( array $input ): array|WP_Error {
-		return ( new WP_Codebox_Artifacts() )->apply_approved( $input );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	public static function stage_artifact_apply( array $input ): array|WP_Error {
-		return WP_Codebox_Data_Machine_Pending_Actions::stage_apply_artifact( $input );
-	}
-
-	public static function can_run_agent_task(): bool {
-		$allowed = current_user_can( 'manage_options' );
-
-		return (bool) apply_filters( 'wp_codebox_can_run_agent_task', $allowed );
-	}
-
-	/** @param mixed $input Ability input or request-like object. */
-	public static function can_create_browser_playground_session( mixed $input = null ): bool {
-		$allowed          = current_user_can( 'manage_options' );
-		$filtered_allowed = (bool) apply_filters( 'wp_codebox_can_run_agent_task', $allowed );
-		if ( $filtered_allowed ) {
-			return true;
-		}
-
-		if ( $allowed ) {
-			return false;
-		}
-
-		$input = self::permission_input_array( $input );
-		if ( empty( $input ) ) {
-			return false;
-		}
-
-		$authorization = self::trusted_orchestrator_authorization( $input, self::BROWSER_SESSION_CREATE_SCOPE );
-
-		return true === ( $authorization['authorized'] ?? false );
-	}
-
-	/** @param mixed $input Ability input or request-like object. */
-	public static function can_persist_browser_artifact( mixed $input = null ): bool {
-		if ( current_user_can( 'manage_options' ) ) {
-			return true;
-		}
-
-		$input = self::permission_input_array( $input );
-		if ( empty( $input ) ) {
-			return false;
-		}
-
-		$authorization = self::trusted_orchestrator_authorization( $input, self::BROWSER_ARTIFACT_WRITE_SCOPE );
-
-		return true === ( $authorization['authorized'] ?? false );
-	}
-
-	/** @param mixed $input Ability input or request-like object. @return array<string,mixed> */
-	private static function permission_input_array( mixed $input ): array {
-		if ( is_array( $input ) ) {
-			return $input;
-		}
-
-		if ( is_object( $input ) && is_callable( array( $input, 'get_json_params' ) ) ) {
-			$params = $input->get_json_params();
-			return is_array( $params ) ? $params : array();
-		}
-
-		if ( is_object( $input ) && is_callable( array( $input, 'get_params' ) ) ) {
-			$params = $input->get_params();
-			return is_array( $params ) ? $params : array();
-		}
-
-		return array();
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed> */
-	private static function browser_session_authorization( array $input ): array {
-		return self::trusted_orchestrator_authorization( $input, self::BROWSER_SESSION_CREATE_SCOPE );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed> */
-	private static function trusted_orchestrator_authorization( array $input, string $required_scope ): array {
-		$authorization = is_array( $input['authorization'] ?? null ) ? $input['authorization'] : array();
-		$caller        = trim( (string) ( $authorization['caller'] ?? '' ) );
-		$scope         = trim( (string) ( $authorization['scope'] ?? '' ) );
-		$result        = array_filter(
-			array(
-				'schema'     => 'wp-codebox/trusted-orchestrator-authorization/v1',
-				'caller'     => $caller,
-				'scope'      => $scope,
-				'authorized' => false,
-				'method'     => 'trusted-orchestrator',
-				'reason'     => 'missing-authorization',
-			),
-			static fn( mixed $value ): bool => '' !== $value
-		);
-
-		if ( '' === $caller ) {
-			return $result;
-		}
-
-		if ( $required_scope !== $scope ) {
-			$result['reason'] = 'missing-scope';
-			return $result;
-		}
-
-		/**
-		 * Filters trusted browser-session callers.
-		 *
-		 * Return either a map of caller ids to scopes, or a list of grant arrays:
-		 * [ 'studio-web' => [ 'browser-session:create' ] ]
-		 * [ [ 'caller' => 'studio-web', 'scopes' => [ 'browser-session:create' ] ] ]
-		 *
-		 * @param array<int|string,mixed> $trusted_callers Trusted caller grants.
-		 * @param array<string,mixed>     $authorization   Explicit caller authorization payload.
-		 * @param array<string,mixed>     $input           Ability input.
-		 */
-		$trusted_callers = apply_filters( 'wp_codebox_trusted_browser_session_callers', array(), $authorization, $input );
-		$trusted_callers = is_array( $trusted_callers ) ? $trusted_callers : array();
-
-		if ( self::trusted_browser_session_caller_has_scope( $trusted_callers, $caller, $scope ) ) {
-			$result['authorized'] = true;
-			$result['reason']     = 'trusted-caller-grant';
-			return $result;
-		}
-
-		$result['reason'] = 'caller-not-trusted';
-
-		return $result;
-	}
-
-	/** @param array<int|string,mixed> $trusted_callers Trusted caller grants. */
-	private static function trusted_browser_session_caller_has_scope( array $trusted_callers, string $caller, string $scope ): bool {
-		foreach ( $trusted_callers as $key => $grant ) {
-			if ( is_string( $key ) && $caller === $key ) {
-				$scopes = is_array( $grant ) ? $grant : array( $grant );
-				return in_array( $scope, array_map( 'strval', $scopes ), true );
-			}
-
-			if ( ! is_array( $grant ) || $caller !== (string) ( $grant['caller'] ?? '' ) ) {
-				continue;
-			}
-
-			$scopes = is_array( $grant['scopes'] ?? null ) ? $grant['scopes'] : array( $grant['scope'] ?? '' );
-			if ( in_array( $scope, array_map( 'strval', $scopes ), true ) ) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	private static function normalize_task_input( array $input ): array|WP_Error {
-		return WP_Codebox_Agent_Task::normalize_input( $input, fn( array $tools ): WP_Error|null => ( new WP_Codebox_Agent_Sandbox_Runner() )->validate_allowed_tools( $tools ), true );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array{connectors:string[],settings:string[]} */
-	private static function browser_inheritance_request( array $input ): array {
-		$inherit = is_array( $input['inherit'] ?? null ) ? $input['inherit'] : array();
-
-		return array(
-			'connectors' => self::string_list( $inherit['connectors'] ?? array() ),
-			'settings'   => self::string_list( $inherit['settings'] ?? array() ),
-		);
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array{inheritance:array{connectors:array<int,array<string,mixed>>,settings:array<int,array<string,mixed>>}}|WP_Error */
-	private static function browser_inheritance_resolution_payload( array $input ): array|WP_Error {
-		$request    = self::browser_inheritance_request( $input );
-		$resolution = array(
-			'connectors' => array_map(
-				static fn( string $name ): array => array(
-					'name'   => $name,
-					'status' => 'unresolved',
-				),
-				$request['connectors']
-			),
-			'settings'   => array_map(
-				static fn( string $name ): array => array(
-					'name'   => $name,
-					'status' => 'unresolved',
-				),
-				$request['settings']
-			),
-		);
-
-		if ( function_exists( 'apply_filters' ) && ( ! empty( $request['connectors'] ) || ! empty( $request['settings'] ) ) ) {
-			$filtered = apply_filters( 'wp_codebox_resolve_inheritance', $resolution, $request, $input );
-			if ( is_array( $filtered ) ) {
-				$resolution = $filtered;
-			}
-		}
-
-		$inheritance = array(
-			'connectors' => self::sanitize_browser_inheritance_connectors( $resolution['connectors'] ?? array() ),
-			'settings'   => self::sanitize_browser_inheritance_settings( $resolution['settings'] ?? array() ),
-		);
-		$credential_error = self::browser_connector_credentials_error( $inheritance );
-		if ( null !== $credential_error ) {
-			return $credential_error;
-		}
-
-		return array( 'inheritance' => $inheritance );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @param array{connectors:array<int,array<string,mixed>>,settings:array<int,array<string,mixed>>} $inheritance @return array<string,mixed> */
-	private static function browser_input_with_inheritance( array $input, array $inheritance ): array {
-		$input['provider_plugin_paths'] = array_values( array_unique( array_merge( self::browser_provider_plugin_paths( $input ), self::browser_inheritance_provider_plugin_paths( $inheritance ) ) ) );
-		$input['secret_env']            = array_values( array_unique( array_merge( self::browser_secret_env_names( $input ), self::browser_inheritance_secret_env_names( $inheritance ) ) ) );
-
-		return $input;
-	}
-
-	/** @param array<string,mixed> $input Ability input. @param array<string,mixed> $task_input Normalized task input. @param array<int,array<string,mixed>> $artifacts Browser artifact specs. @param array{connectors:array<int,array<string,mixed>>,settings:array<int,array<string,mixed>>} $inheritance @return array<string,mixed> */
-	private static function browser_task_payload( array $input, array $task_input, string $session_id, array $artifacts, array $inheritance ): array {
-		return array_filter(
-			array(
-				'schema'      => 'wp-codebox/browser-agent-task-payload/v1',
-				'agent'       => self::browser_agent_slug( $input ),
-				'mode'        => self::browser_mode( $input ),
-				'provider'    => self::browser_provider( $input, $inheritance ),
-				'model'       => self::browser_model( $input, $inheritance ),
-				'message'     => (string) $task_input['goal'],
-				'session_id'  => $session_id,
-				'task_input'  => $task_input,
-				'agent_bundles' => self::normalize_agent_bundles( $input['agent_bundles'] ?? $input['agentBundles'] ?? array() ),
-				'inheritance' => $inheritance,
-				'secret_env'  => self::browser_secret_env_names( $input ),
-				'artifacts'   => array(
-					'schema' => 'wp-codebox/browser-artifacts/v1',
-					'files'  => $artifacts,
-				),
-			),
-			static fn( mixed $value ): bool => '' !== $value && array() !== $value
-		);
-	}
-
-	/** @param array<string,mixed> $input Ability input. */
-	private static function browser_agent_slug( array $input ): string {
-		$agent = trim( (string) ( $input['agent'] ?? '' ) );
-		if ( '' === $agent && function_exists( 'apply_filters' ) ) {
-			$agent = (string) apply_filters( 'wp_codebox_default_agent', '' );
-		}
-
-		return '' !== trim( $agent ) ? trim( $agent ) : 'wp-codebox-sandbox';
-	}
-
-	/** @param array<string,mixed> $input Ability input. */
-	private static function browser_mode( array $input ): string {
-		$mode = trim( (string) ( $input['mode'] ?? '' ) );
-		return '' !== $mode ? $mode : 'sandbox';
-	}
-
-	/** @param array<string,mixed> $input Ability input. @param array{connectors:array<int,array<string,mixed>>,settings:array<int,array<string,mixed>>} $inheritance */
-	private static function browser_provider( array $input, array $inheritance ): string {
-		$provider = trim( (string) ( $input['provider'] ?? '' ) );
-		if ( '' !== $provider ) {
-			return $provider;
-		}
-
-		foreach ( $inheritance['connectors'] as $connector ) {
-			$provider = trim( (string) ( $connector['provider'] ?? '' ) );
-			if ( '' !== $provider ) {
-				return $provider;
-			}
-		}
-
-		return function_exists( 'apply_filters' ) ? trim( (string) apply_filters( 'wp_codebox_default_provider', '' ) ) : '';
-	}
-
-	/** @param array<string,mixed> $input Ability input. @param array{connectors:array<int,array<string,mixed>>,settings:array<int,array<string,mixed>>} $inheritance */
-	private static function browser_model( array $input, array $inheritance ): string {
-		$model = trim( (string) ( $input['model'] ?? '' ) );
-		if ( '' !== $model ) {
-			return $model;
-		}
-
-		foreach ( $inheritance['connectors'] as $connector ) {
-			$model = trim( (string) ( $connector['model'] ?? '' ) );
-			if ( '' !== $model ) {
-				return $model;
-			}
-		}
-
-		return function_exists( 'apply_filters' ) ? trim( (string) apply_filters( 'wp_codebox_default_model', '' ) ) : '';
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return string[] */
-	private static function browser_provider_plugin_paths( array $input ): array {
-		return array_values( array_unique( array_filter( array_map( static fn( $path ): string => self::browser_clean_path( (string) $path ), is_array( $input['provider_plugin_paths'] ?? null ) ? $input['provider_plugin_paths'] : array() ), static fn( string $path ): bool => '' !== $path && is_dir( $path ) ) ) );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return string[] */
-	private static function browser_secret_env_names( array $input ): array {
-		return array_values( array_unique( array_filter( self::string_list( $input['secret_env'] ?? array() ), static fn( string $name ): bool => 1 === preg_match( '/^[A-Z_][A-Z0-9_]*$/', $name ) ) ) );
-	}
-
-	/** @return array<int,array<string,mixed>> */
-	private static function normalize_agent_bundles( mixed $bundles ): array {
-		$normalized = array();
-		foreach ( is_array( $bundles ) ? $bundles : array() as $bundle ) {
-			if ( ! is_array( $bundle ) ) {
-				continue;
-			}
-			$source = isset( $bundle['source'] ) ? trim( (string) $bundle['source'] ) : '';
-			$inline = is_array( $bundle['bundle'] ?? null ) ? $bundle['bundle'] : null;
-			if ( '' === $source && null === $inline ) {
-				continue;
-			}
-
-			$entry = array();
-			if ( '' !== $source ) {
-				$entry['source'] = $source;
-			}
-			if ( null !== $inline ) {
-				$entry['bundle'] = $inline;
-			}
-			foreach ( array( 'slug', 'token_env' ) as $field ) {
-				$value = isset( $bundle[ $field ] ) ? trim( (string) $bundle[ $field ] ) : '';
-				if ( '' !== $value ) {
-					$entry[ $field ] = $value;
-				}
-			}
-			$on_conflict = (string) ( $bundle['on_conflict'] ?? 'upgrade' );
-			$entry['on_conflict'] = in_array( $on_conflict, array( 'error', 'skip', 'upgrade' ), true ) ? $on_conflict : 'upgrade';
-			if ( isset( $bundle['owner_id'] ) && (int) $bundle['owner_id'] > 0 ) {
-				$entry['owner_id'] = (int) $bundle['owner_id'];
-			}
-
-			$normalized[] = $entry;
-		}
-
-		return $normalized;
-	}
-
-	/** @param array{connectors:array<int,array<string,mixed>>,settings:array<int,array<string,mixed>>} $inheritance @return string[] */
-	private static function browser_inheritance_provider_plugin_paths( array $inheritance ): array {
-		$paths = array();
-		foreach ( $inheritance['connectors'] as $connector ) {
-			$paths = array_merge( $paths, self::string_list( $connector['providerPluginPaths'] ?? array() ) );
-		}
-
-		return array_values( array_unique( $paths ) );
-	}
-
-	/** @param array{connectors:array<int,array<string,mixed>>,settings:array<int,array<string,mixed>>} $inheritance @return string[] */
-	private static function browser_inheritance_secret_env_names( array $inheritance ): array {
-		$names = array();
-		foreach ( $inheritance['connectors'] as $connector ) {
-			$names = array_merge( $names, self::string_list( $connector['secretEnv'] ?? array() ) );
-			$credentials = is_array( $connector['credentials'] ?? null ) ? $connector['credentials'] : array();
-			foreach ( is_array( $credentials['secrets'] ?? null ) ? $credentials['secrets'] : array() as $secret ) {
-				if ( is_array( $secret ) && 'available' === ( $secret['status'] ?? '' ) ) {
-					$names[] = (string) ( $secret['name'] ?? '' );
-				}
-			}
-		}
-
-		return array_values( array_unique( array_filter( $names ) ) );
-	}
-
-	/** @param array<int,mixed> $connectors Inheritance connector rows. @return array<int,array<string,mixed>> */
-	private static function sanitize_browser_inheritance_connectors( array $connectors ): array {
-		$sanitized = array();
-		foreach ( $connectors as $connector ) {
-			if ( ! is_array( $connector ) ) {
-				continue;
-			}
-
-			$name = trim( (string) ( $connector['name'] ?? '' ) );
-			if ( '' === $name ) {
-				continue;
-			}
-
-			$entry = array(
-				'name'   => $name,
-				'status' => trim( (string) ( $connector['status'] ?? 'resolved' ) ),
-			);
-			foreach ( array( 'provider', 'model' ) as $field ) {
-				$value = trim( (string) ( $connector[ $field ] ?? '' ) );
-				if ( '' !== $value ) {
-					$entry[ $field ] = $value;
-				}
-			}
-
-			$provider_plugin_paths = array_values( array_filter( array_map( static fn( string $path ): string => self::browser_clean_path( $path ), self::string_list( $connector['provider_plugin_paths'] ?? $connector['providerPluginPaths'] ?? array() ) ), static fn( string $path ): bool => '' !== $path && is_dir( $path ) ) );
-			if ( ! empty( $provider_plugin_paths ) ) {
-				$entry['providerPluginPaths'] = array_values( array_unique( $provider_plugin_paths ) );
-			}
-
-			$secret_env = array_values( array_filter( self::string_list( $connector['secret_env'] ?? $connector['secretEnv'] ?? array() ), static fn( string $name ): bool => 1 === preg_match( '/^[A-Z_][A-Z0-9_]*$/', $name ) ) );
-			if ( ! empty( $secret_env ) ) {
-				$entry['secretEnv'] = array_values( array_unique( $secret_env ) );
-			}
-
-			$credentials = self::sanitize_browser_connector_credentials( $connector['credentials'] ?? null, $name );
-			if ( ! empty( $credentials ) ) {
-				$entry['credentials'] = $credentials;
-			}
-
-			$sanitized[] = $entry;
-		}
-
-		return $sanitized;
-	}
-
-	/** @return array<string,mixed> */
-	private static function sanitize_browser_connector_credentials( mixed $credentials, string $connector_name ): array {
-		if ( ! is_array( $credentials ) ) {
-			return array();
-		}
-
-		$status = self::browser_credential_status( (string) ( $credentials['status'] ?? 'missing' ) );
-		$entry  = array(
-			'schema'    => 'wp-codebox/connector-credentials/v1',
-			'connector' => $connector_name,
-			'scope'     => 'connector',
-			'status'    => $status,
-			'secrets'   => array(),
-		);
-		$reason = self::browser_redacted_reason( $credentials['reason'] ?? '' );
-		if ( '' !== $reason ) {
-			$entry['reason'] = $reason;
-		}
-
-		foreach ( is_array( $credentials['secrets'] ?? null ) ? $credentials['secrets'] : array() as $secret ) {
-			if ( ! is_array( $secret ) ) {
-				continue;
-			}
-
-			$name = trim( (string) ( $secret['name'] ?? '' ) );
-			if ( 1 !== preg_match( '/^[A-Z_][A-Z0-9_]*$/', $name ) ) {
-				continue;
-			}
-
-			$secret_entry = array(
-				'name'   => $name,
-				'status' => self::browser_credential_status( (string) ( $secret['status'] ?? $status ) ),
-			);
-			foreach ( array( 'scope', 'source', 'reason' ) as $field ) {
-				$value = 'reason' === $field ? self::browser_redacted_reason( $secret[ $field ] ?? '' ) : trim( (string) ( $secret[ $field ] ?? '' ) );
-				if ( '' !== $value ) {
-					$secret_entry[ $field ] = $value;
-				}
-			}
-
-			$entry['secrets'][] = $secret_entry;
-		}
-
-		return $entry;
-	}
-
-	private static function browser_credential_status( string $status ): string {
-		return in_array( $status, array( 'available', 'missing', 'denied' ), true ) ? $status : 'missing';
-	}
-
-	private static function browser_redacted_reason( mixed $reason ): string {
-		$reason = trim( (string) $reason );
-		return '' === $reason ? '' : substr( preg_replace( '/[^A-Za-z0-9 .:_-]/', '', $reason ) ?? '', 0, 160 );
-	}
-
-	/** @param array{connectors:array<int,array<string,mixed>>,settings:array<int,array<string,mixed>>} $inheritance */
-	private static function browser_connector_credentials_error( array $inheritance ): WP_Error|null {
-		$failures = array();
-		foreach ( $inheritance['connectors'] as $connector ) {
-			$credentials = is_array( $connector['credentials'] ?? null ) ? $connector['credentials'] : array();
-			if ( empty( $credentials ) ) {
-				continue;
-			}
-
-			$status  = (string) ( $credentials['status'] ?? 'missing' );
-			$secrets = array_filter( is_array( $credentials['secrets'] ?? null ) ? $credentials['secrets'] : array(), static fn( mixed $secret ): bool => is_array( $secret ) && in_array( (string) ( $secret['status'] ?? '' ), array( 'missing', 'denied' ), true ) );
-			if ( in_array( $status, array( 'missing', 'denied' ), true ) || ! empty( $secrets ) ) {
-				$failures[] = array(
-					'name'        => (string) ( $connector['name'] ?? '' ),
-					'status'      => (string) ( $connector['status'] ?? '' ),
-					'credentials' => $credentials,
-				);
-			}
-		}
-
-		return empty( $failures ) ? null : new WP_Error( 'wp_codebox_connector_credentials_unavailable', 'Requested connector credentials are missing or denied for this browser sandbox scope.', array( 'status' => 403, 'schema' => 'wp-codebox/connector-credential-failure/v1', 'connectors' => $failures ) );
-	}
-
-	/** @param array<int,mixed> $settings Inheritance setting rows. @return array<int,array<string,mixed>> */
-	private static function sanitize_browser_inheritance_settings( array $settings ): array {
-		$sanitized = array();
-		foreach ( $settings as $setting ) {
-			if ( ! is_array( $setting ) ) {
-				continue;
-			}
-			$name = trim( (string) ( $setting['name'] ?? '' ) );
-			if ( '' === $name ) {
-				continue;
-			}
-			$entry = array(
-				'name'   => $name,
-				'status' => trim( (string) ( $setting['status'] ?? 'resolved' ) ),
-			);
-			$scope = trim( (string) ( $setting['scope'] ?? '' ) );
-			if ( '' !== $scope ) {
-				$entry['scope'] = $scope;
-			}
-			$sanitized[] = $entry;
-		}
-
-		return $sanitized;
-	}
-
-	/** @return string[] */
-	private static function string_list( mixed $value ): array {
-		return WP_Codebox_Agent_Task::string_list( $value );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<int,array<string,mixed>>|WP_Error */
-	private static function browser_artifact_files( array $input ): array|WP_Error {
-		$files      = is_array( $input['artifact_files'] ?? null ) ? $input['artifact_files'] : array();
-		$playground = is_array( $input['playground'] ?? null ) ? $input['playground'] : array();
-		$base_path  = self::browser_artifact_base_path( $playground );
-		$base_url   = self::browser_artifact_base_url( $playground );
-		$normalized = array();
-		foreach ( $files as $index => $file ) {
-			if ( ! is_array( $file ) ) {
-				return new WP_Error( 'wp_codebox_browser_artifact_file_invalid', 'Each browser artifact file must be an object.', array( 'status' => 400, 'index' => $index ) );
-			}
-
-			$path_validation = self::validate_browser_artifact_path( trim( (string) ( $file['path'] ?? '' ) ), (int) $index );
-			if ( is_wp_error( $path_validation ) ) {
-				return $path_validation;
-			}
-			$path = $path_validation;
-
-			$encoding = strtolower( trim( (string) ( $file['encoding'] ?? '' ) ) );
-			if ( '' === $encoding ) {
-				$encoding = array_key_exists( 'content_base64', $file ) ? 'base64' : 'utf-8';
-			}
-
-			if ( ! in_array( $encoding, array( 'utf-8', 'base64' ), true ) ) {
-				return new WP_Error( 'wp_codebox_browser_artifact_encoding_invalid', 'Browser artifact file encoding must be utf-8 or base64.', array( 'status' => 400, 'index' => $index, 'path' => $path, 'encoding' => $encoding ) );
-			}
-
-			if ( 'base64' === $encoding ) {
-				$encoded = (string) ( $file['content_base64'] ?? $file['content'] ?? '' );
-				$bytes   = base64_decode( $encoded, true );
-				if ( false === $bytes ) {
-					return new WP_Error( 'wp_codebox_browser_artifact_base64_invalid', 'Browser artifact file content_base64 must be valid base64.', array( 'status' => 400, 'index' => $index, 'path' => $path ) );
-				}
-			} else {
-				$bytes   = (string) ( $file['content'] ?? '' );
-				$encoded = '';
-			}
-
-			$size = strlen( $bytes );
-			if ( $size > self::BROWSER_ARTIFACT_MAX_BYTES ) {
-				return new WP_Error( 'wp_codebox_browser_artifact_file_too_large', 'Browser artifact file exceeds the maximum inline size.', array( 'status' => 400, 'index' => $index, 'path' => $path, 'size' => $size, 'max_size' => self::BROWSER_ARTIFACT_MAX_BYTES ) );
-			}
-
-			$mime_type = trim( (string) ( $file['mime_type'] ?? '' ) );
-			if ( '' === $mime_type ) {
-				$mime_type = self::browser_artifact_mime_type( $path );
-			}
-
-			$artifact = array(
-				'path'            => $path,
-				'playground_path' => self::join_browser_path( $base_path, $path ),
-				'url_path'        => self::join_browser_path( $base_url, $path ),
-				'encoding'        => $encoding,
-				'mime_type'       => $mime_type,
-				'size'            => $size,
-				'sha256'          => hash( 'sha256', $bytes ),
-				'kind'            => (string) ( $file['kind'] ?? 'text' ),
-				'description'     => (string) ( $file['description'] ?? '' ),
-			);
-
-			if ( 'base64' === $encoding ) {
-				$artifact['content_base64'] = $encoded;
-			} else {
-				$artifact['content'] = $bytes;
-			}
-
-			$normalized[] = $artifact;
-		}
-
-		return $normalized;
-	}
-
-	private static function validate_browser_artifact_path( string $path, int $index ): string|WP_Error {
-		if ( '' === $path || str_starts_with( $path, '/' ) || ! preg_match( '#^[A-Za-z0-9_./-]+$#', $path ) ) {
-			return new WP_Error( 'wp_codebox_browser_artifact_path_invalid', 'Browser artifact file paths must be safe relative paths.', array( 'status' => 400, 'index' => $index, 'path' => $path ) );
-		}
-
-		$segments = explode( '/', $path );
-		foreach ( $segments as $segment ) {
-			if ( '' === $segment || '.' === $segment || '..' === $segment ) {
-				return new WP_Error( 'wp_codebox_browser_artifact_path_invalid', 'Browser artifact file paths must not contain empty, current-directory, or parent-directory segments.', array( 'status' => 400, 'index' => $index, 'path' => $path, 'segment' => $segment ) );
-			}
-		}
-
-		$extension = strtolower( pathinfo( $path, PATHINFO_EXTENSION ) );
-		if ( in_array( $extension, array( 'php', 'phtml', 'phar', 'cgi', 'pl', 'py', 'rb', 'asp', 'aspx', 'jsp' ), true ) ) {
-			return new WP_Error( 'wp_codebox_browser_artifact_extension_blocked', 'Browser artifact files must not use executable server-side extensions.', array( 'status' => 400, 'index' => $index, 'path' => $path, 'extension' => $extension ) );
-		}
-
-		return $path;
-	}
-
-	private static function browser_artifact_mime_type( string $path ): string {
-		return match ( strtolower( pathinfo( $path, PATHINFO_EXTENSION ) ) ) {
-			'html', 'htm' => 'text/html',
-			'css'        => 'text/css',
-			'js', 'mjs'   => 'text/javascript',
-			'json'       => 'application/json',
-			'svg'        => 'image/svg+xml',
-			'jpg', 'jpeg' => 'image/jpeg',
-			'png'        => 'image/png',
-			'webp'       => 'image/webp',
-			'gif'        => 'image/gif',
-			'txt'        => 'text/plain',
-			default      => 'application/octet-stream',
-		};
-	}
-
-	/** @param array<int,mixed> $mu_plugins Mu-plugin dependency specs. @return array<int,array<string,mixed>>|WP_Error */
-	private static function normalize_browser_mu_plugins( array $mu_plugins ): array|WP_Error {
-		$normalized = array();
-		foreach ( $mu_plugins as $index => $mu_plugin ) {
-			if ( ! is_array( $mu_plugin ) ) {
-				return new WP_Error( 'wp_codebox_browser_mu_plugin_invalid', 'Each browser mu-plugin must be an object.', array( 'status' => 400, 'index' => $index ) );
-			}
-
-			$slug = self::safe_key( (string) ( $mu_plugin['slug'] ?? '' ) );
-			$file = trim( (string) ( $mu_plugin['file'] ?? ( '' !== $slug ? $slug . '.php' : '' ) ) );
-			if ( '' === $file || str_contains( $file, '..' ) || str_contains( $file, '/' ) || ! str_ends_with( $file, '.php' ) || ! preg_match( '#^[A-Za-z0-9_.-]+$#', $file ) ) {
-				return new WP_Error( 'wp_codebox_browser_mu_plugin_file_invalid', 'Browser mu-plugin files must be safe PHP filenames.', array( 'status' => 400, 'index' => $index ) );
-			}
-
-			$source_path = self::browser_clean_path( (string) ( $mu_plugin['path'] ?? '' ) );
-			$source_url  = trim( (string) ( $mu_plugin['url'] ?? '' ) );
-			$package     = null;
-			$provenance  = array();
-			if ( '' !== $source_path || '' !== $source_url ) {
-				if ( '' === $slug ) {
-					return new WP_Error( 'wp_codebox_browser_mu_plugin_slug_missing', 'Packaged browser mu-plugin specs require a slug.', array( 'status' => 400, 'index' => $index ) );
-				}
-
-				if ( '' !== $source_path ) {
-					if ( ! is_dir( $source_path ) ) {
-						return new WP_Error( 'wp_codebox_browser_mu_plugin_path_missing', 'Browser mu-plugin path does not exist.', array( 'status' => 400, 'index' => $index, 'slug' => $slug ) );
-					}
-
-					$package    = self::browser_package_component_plugin( $slug, $source_path );
-					$provenance = array(
-						'schema' => 'wp-codebox/browser-mu-plugin-provenance/v1',
-						'source' => 'runtime-mu-plugin-path',
-						'path'   => $source_path,
-					);
-				} else {
-					$package    = self::browser_remote_mu_plugin_package( $slug, $source_url, $index, (string) ( $mu_plugin['sha256'] ?? '' ) );
-					$provenance = array(
-						'schema' => 'wp-codebox/browser-mu-plugin-provenance/v1',
-						'source' => 'runtime-mu-plugin-remote-package',
-						'url'    => $source_url,
-					);
-				}
-
-				if ( is_wp_error( $package ) ) {
-					return $package;
-				}
-			}
-
-			$content = (string) ( $mu_plugin['content'] ?? '' );
-			if ( null === $package && '' === trim( $content ) ) {
-				return new WP_Error( 'wp_codebox_browser_mu_plugin_content_missing', 'Browser mu-plugin content is required.', array( 'status' => 400, 'index' => $index ) );
-			}
-
-			$target_folder = sanitize_key( (string) ( $mu_plugin['targetFolderName'] ?? $slug ) );
-			$entry         = ltrim( str_replace( '\\', '/', (string) ( $mu_plugin['entry'] ?? ( '' !== $target_folder ? $target_folder . '.php' : '' ) ) ), '/' );
-			if ( null !== $package && ( '' === $target_folder || '' === $entry || str_contains( $entry, '..' ) || str_starts_with( $entry, '/' ) || ! str_ends_with( $entry, '.php' ) ) ) {
-				return new WP_Error( 'wp_codebox_browser_mu_plugin_entry_invalid', 'Packaged browser mu-plugins require a safe PHP entry file.', array( 'status' => 400, 'index' => $index, 'slug' => $slug ) );
-			}
-
-			$normalized[] = array_filter( array(
-				'slug'            => '' !== $slug ? $slug : self::safe_key( basename( $file, '.php' ) ),
-				'file'            => $file,
-				'path'            => '/wordpress/wp-content/mu-plugins/' . $file,
-				'content'         => $content,
-				'url'             => is_array( $package ) ? $package['url'] : '',
-				'sha256'          => is_array( $package ) ? $package['sha256'] : '',
-				'local_package_fetch_url' => is_array( $package ) ? $package['fetch_url'] : '',
-				'targetFolderName' => $target_folder,
-				'entry'           => $entry,
-				'local_package'   => null !== $package,
-				'provenance'      => array_filter( array_merge( $provenance, array( 'sha256' => is_array( $package ) ? $package['sha256'] : '' ) ) ),
-				'readiness'       => 'compiled',
-			), static fn( mixed $value ): bool => null !== $value && '' !== $value && array() !== $value );
-		}
-
-		return $normalized;
-	}
-
-	/** @return array{url:string,fetch_url:string,path:string,sha256:string} | WP_Error */
-	private static function browser_remote_mu_plugin_package( string $slug, string $url, int $index, string $expected_sha256 = '' ): array|WP_Error {
-		$sha256 = strtolower( trim( $expected_sha256 ) );
-		if ( '' !== $sha256 && ! preg_match( '/^[a-f0-9]{64}$/', $sha256 ) ) {
-			return new WP_Error( 'wp_codebox_browser_mu_plugin_sha256_invalid', 'Browser mu-plugin sha256 must be a 64-character hex digest.', array( 'status' => 400, 'index' => $index, 'slug' => $slug ) );
-		}
-
-		return self::browser_package_remote_plugin( $slug, $url, $index, $sha256 );
-	}
-
-	/** @param array<int,mixed> $themes Theme dependency specs. @return array<int,array<string,mixed>>|WP_Error */
-	private static function normalize_browser_themes( array $themes ): array|WP_Error {
-		$normalized = array();
-		foreach ( $themes as $index => $theme ) {
-			if ( ! is_array( $theme ) ) {
-				return new WP_Error( 'wp_codebox_browser_theme_invalid', 'Each browser theme must be an object.', array( 'status' => 400, 'index' => $index ) );
-			}
-
-			$slug  = self::safe_key( (string) ( $theme['slug'] ?? '' ) );
-			$url   = trim( (string) ( $theme['url'] ?? '' ) );
-			$files = is_array( $theme['files'] ?? null ) ? $theme['files'] : array();
-			if ( '' === $slug ) {
-				return new WP_Error( 'wp_codebox_browser_theme_slug_missing', 'Browser theme slug is required.', array( 'status' => 400, 'index' => $index ) );
-			}
-			if ( '' === $url && empty( $files ) ) {
-				return new WP_Error( 'wp_codebox_browser_theme_source_missing', 'Browser themes require a zip URL or files.', array( 'status' => 400, 'index' => $index ) );
-			}
-
-			$source = null;
-			if ( '' !== $url ) {
-				$source = self::browser_theme_url( $url, $index );
-				if ( is_wp_error( $source ) ) {
-					return $source;
-				}
-			}
-
-			$normalized_files = self::normalize_browser_theme_files( $files, $slug, $index );
-			if ( is_wp_error( $normalized_files ) ) {
-				return $normalized_files;
-			}
-
-			$normalized[] = array_filter(
-				array(
-					'slug'       => $slug,
-					'url'        => $source['url'] ?? '',
-					'activate'   => ! array_key_exists( 'activate', $theme ) || (bool) $theme['activate'],
-					'files'      => $normalized_files,
-					'readiness'  => 'compiled',
-					'provenance' => $source ? array_filter( array( 'schema' => 'wp-codebox/browser-theme-provenance/v1', 'url' => $source['url'], 'origin' => $source['origin'], 'host' => $source['host'] ) ) : array(),
-				),
-				static fn( mixed $value ): bool => array() !== $value && '' !== $value
-			);
-		}
-
-		return $normalized;
-	}
-
-	/** @param array<int,mixed> $files Theme file specs. @return array<int,array<string,string>>|WP_Error */
-	private static function normalize_browser_theme_files( array $files, string $slug, int $theme_index ): array|WP_Error {
-		$normalized = array();
-		foreach ( $files as $index => $file ) {
-			if ( ! is_array( $file ) ) {
-				return new WP_Error( 'wp_codebox_browser_theme_file_invalid', 'Each browser theme file must be an object.', array( 'status' => 400, 'theme_index' => $theme_index, 'index' => $index ) );
-			}
-
-			$path = trim( (string) ( $file['path'] ?? '' ) );
-			if ( '' === $path || str_contains( $path, '..' ) || str_starts_with( $path, '/' ) || ! preg_match( '#^[A-Za-z0-9_./-]+$#', $path ) ) {
-				return new WP_Error( 'wp_codebox_browser_theme_file_path_invalid', 'Browser theme file paths must be safe relative paths.', array( 'status' => 400, 'theme_index' => $theme_index, 'index' => $index ) );
-			}
-
-			$normalized[] = array(
-				'path'            => $path,
-				'playground_path' => '/wordpress/wp-content/themes/' . $slug . '/' . $path,
-				'content'         => (string) ( $file['content'] ?? '' ),
-			);
-		}
-
-		return $normalized;
-	}
-
-	/** @param array<int,mixed> $operations Bootstrap operation specs. @return array<int,array<string,mixed>>|WP_Error */
-	private static function normalize_browser_bootstrap( array $operations ): array|WP_Error {
-		$normalized = array();
-		foreach ( $operations as $index => $operation ) {
-			if ( ! is_array( $operation ) ) {
-				return new WP_Error( 'wp_codebox_browser_bootstrap_invalid', 'Each browser bootstrap operation must be an object.', array( 'status' => 400, 'index' => $index ) );
-			}
-
-			$name = self::safe_key( (string) ( $operation['operation'] ?? $operation['name'] ?? '' ) );
-			if ( ! in_array( $name, array( 'set_option', 'activate_plugin', 'activate_theme', 'flush_rewrite_rules' ), true ) ) {
-				return new WP_Error( 'wp_codebox_browser_bootstrap_operation_invalid', 'Browser bootstrap operation is not supported.', array( 'status' => 400, 'index' => $index, 'operation' => $name ) );
-			}
-
-			$normalized[] = array(
-				'operation' => $name,
-				'args'      => is_array( $operation['args'] ?? null ) ? $operation['args'] : array_diff_key( $operation, array( 'operation' => true, 'name' => true ) ),
-				'readiness' => 'compiled',
-			);
-		}
-
-		return $normalized;
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	private static function browser_playground( array $input ): array|WP_Error {
-		$playground = is_array( $input['playground'] ?? null ) ? $input['playground'] : array();
-		$client     = self::browser_trusted_url(
-			(string) ( $playground['client_module_url'] ?? 'https://playground.automattic.ai/client/index.js' ),
-			'client_module_url',
-			'wp_codebox_browser_playground_allowed_origins',
-			array( 'https://playground.automattic.ai' )
-		);
-		if ( is_wp_error( $client ) ) {
-			return $client;
-		}
-
-		$remote = self::browser_trusted_url(
-			(string) ( $playground['remote_url'] ?? 'https://playground.automattic.ai/remote.html' ),
-			'remote_url',
-			'wp_codebox_browser_playground_allowed_origins',
-			array( 'https://playground.automattic.ai' )
-		);
-		if ( is_wp_error( $remote ) ) {
-			return $remote;
-		}
-
-		$default_cors_proxy_url = (string) apply_filters( 'wp_codebox_browser_playground_default_cors_proxy_url', 'https://wordpress-playground-cors-proxy.net/?', $playground );
-		$cors_proxy_url         = (string) ( $playground['cors_proxy_url'] ?? $playground['corsProxy'] ?? $default_cors_proxy_url );
-		$cors_proxy             = array( 'url' => '', 'origin' => '', 'host' => '' );
-		if ( '' !== trim( $cors_proxy_url ) ) {
-			$cors_proxy = self::browser_trusted_url(
-				$cors_proxy_url,
-				'cors_proxy_url',
-				'wp_codebox_browser_playground_cors_proxy_allowed_origins',
-				array( 'https://wordpress-playground-cors-proxy.net', 'https://playground.wordpress.net', 'https://playground.automattic.ai' )
-			);
-			if ( is_wp_error( $cors_proxy ) ) {
-				return $cors_proxy;
-			}
-		}
-
-		$playground['client_module_url'] = $client['url'];
-		$playground['remote_url']        = $remote['url'];
-		$playground['cors_proxy_url']    = $cors_proxy['url'];
-		$playground['provenance']        = array(
-			'schema'            => 'wp-codebox/browser-playground-provenance/v1',
-			'client_module_url' => $client,
-			'remote_url'        => $remote,
-			'cors_proxy_url'    => $cors_proxy,
-		);
-
-		return $playground;
-	}
-
-	/** @return array{url:string,origin:string,host:string} | WP_Error */
-	private static function browser_trusted_url( string $url, string $field, string $filter, array $default_allowed_origins ): array|WP_Error {
-		$parts = wp_parse_url( $url );
-		if ( ! is_array( $parts ) || empty( $parts['scheme'] ) || empty( $parts['host'] ) ) {
-			return new WP_Error( 'wp_codebox_browser_url_invalid', 'Browser Playground URL must be absolute.', array( 'status' => 400, 'field' => $field ) );
-		}
-
-		$scheme = strtolower( (string) $parts['scheme'] );
-		if ( 'https' !== $scheme ) {
-			return new WP_Error( 'wp_codebox_browser_url_insecure', 'Browser Playground URL must use https://.', array( 'status' => 400, 'field' => $field ) );
-		}
-
-		$origin  = self::url_origin( $parts );
-		$allowed = self::normalized_origins( apply_filters( $filter, $default_allowed_origins, $field, $url ) );
-		if ( ! in_array( $origin, $allowed, true ) ) {
-			return new WP_Error( 'wp_codebox_browser_origin_not_allowed', 'Browser Playground URL origin is not allowed.', array( 'status' => 400, 'field' => $field, 'origin' => $origin ) );
-		}
-
-		return array(
-			'url'    => $url,
-			'origin' => $origin,
-			'host'   => strtolower( (string) $parts['host'] ),
-		);
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<int,array<string,mixed>>|WP_Error */
-	private static function browser_plugins( array $input ): array|WP_Error {
-		$plugins = is_array( $input['browser_plugins'] ?? null ) ? $input['browser_plugins'] : array();
-		return self::normalize_browser_plugins( $plugins, 'browser_plugins' );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @param array<int,array<string,mixed>> $legacy_plugins Legacy browser_plugins specs. @return array<string,mixed>|WP_Error */
-	private static function browser_runtime_dependencies( array $input, array $legacy_plugins ): array|WP_Error {
-		$runtime = is_array( $input['runtime'] ?? null ) ? $input['runtime'] : array();
-		$runtime_plugin_specs = self::browser_runtime_plugin_specs( array_merge( self::browser_provider_plugin_specs( $input ), is_array( $runtime['plugins'] ?? null ) ? $runtime['plugins'] : array() ) );
-		if ( is_wp_error( $runtime_plugin_specs ) ) {
-			return $runtime_plugin_specs;
-		}
-
-		$runtime_plugins = self::normalize_browser_plugins( $runtime_plugin_specs, 'runtime.plugins' );
-		if ( is_wp_error( $runtime_plugins ) ) {
-			return $runtime_plugins;
-		}
-
-		$mu_plugins = self::normalize_browser_mu_plugins( is_array( $runtime['mu_plugins'] ?? null ) ? $runtime['mu_plugins'] : array() );
-		if ( is_wp_error( $mu_plugins ) ) {
-			return $mu_plugins;
-		}
-
-		$themes = self::normalize_browser_themes( is_array( $runtime['themes'] ?? null ) ? $runtime['themes'] : array() );
-		if ( is_wp_error( $themes ) ) {
-			return $themes;
-		}
-
-		$bootstrap = self::normalize_browser_bootstrap( is_array( $runtime['bootstrap'] ?? null ) ? $runtime['bootstrap'] : array() );
-		if ( is_wp_error( $bootstrap ) ) {
-			return $bootstrap;
-		}
-
-		$component_plugins = self::browser_component_plugins( $input, array_merge( $legacy_plugins, $runtime_plugins ), is_array( $runtime['components'] ?? null ) ? $runtime['components'] : array() );
-		if ( is_wp_error( $component_plugins ) ) {
-			return $component_plugins;
-		}
-
-		$plugins = self::dedupe_browser_plugins( array_merge( $legacy_plugins, $component_plugins, $runtime_plugins ) );
-
-		return array(
-			'schema'                 => 'wp-codebox/browser-runtime-dependencies/v1',
-			'plugins'                => $plugins,
-			'mu_plugins'             => $mu_plugins,
-			'themes'                 => $themes,
-			'bootstrap'              => $bootstrap,
-			'component_plugins'      => count( $component_plugins ),
-			'legacy_browser_plugins' => count( $legacy_plugins ),
-			'summary'                => array(
-				'plugins'    => count( $plugins ),
-				'mu_plugins' => count( $mu_plugins ),
-				'themes'     => count( $themes ),
-				'bootstrap'  => count( $bootstrap ),
-			),
-		);
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<int,array<string,mixed>> */
-	private static function browser_provider_plugin_specs( array $input ): array {
-		return array_map(
-			static fn( string $path ): array => array(
-				'slug'     => self::safe_key( basename( $path ) ),
-				'path'     => $path,
-				'activate' => true,
-				'provenance' => array(
-					'source' => 'provider-plugin-path',
-				),
-			),
-			self::browser_provider_plugin_paths( $input )
-		);
-	}
-
-	/** @param array<int,mixed> $plugins Runtime plugin specs. @return array<int,array<string,mixed>>|WP_Error */
-	private static function browser_runtime_plugin_specs( array $plugins ): array|WP_Error {
-		$resolved = array();
-
-		foreach ( $plugins as $index => $plugin ) {
-			if ( ! is_array( $plugin ) ) {
-				return new WP_Error( 'wp_codebox_browser_plugin_invalid', 'Each browser plugin must be an object.', array( 'status' => 400, 'field' => 'runtime.plugins', 'index' => $index ) );
-			}
-
-			$resource = (string) ( $plugin['resource'] ?? 'url' );
-			$path     = 'git:directory' === $resource ? '' : self::browser_clean_path( (string) ( $plugin['path'] ?? '' ) );
-			$package  = (string) ( $plugin['package'] ?? '' );
-			if ( 'url' === $resource && '' === $path && 'browser' !== $package ) {
-				$slug = self::safe_key( (string) ( $plugin['slug'] ?? '' ) );
-				if ( '' === $slug ) {
-					return new WP_Error( 'wp_codebox_browser_plugin_slug_missing', 'Packaged browser runtime plugin specs require a slug.', array( 'status' => 400, 'field' => 'runtime.plugins', 'index' => $index ) );
-				}
-
-				$package = self::browser_package_remote_plugin( $slug, (string) ( $plugin['url'] ?? '' ), $index, (string) ( $plugin['sha256'] ?? '' ) );
-				if ( is_wp_error( $package ) ) {
-					return $package;
-				}
-
-				$resolved[] = array_merge(
-					$plugin,
-					array(
-						'slug'          => $slug,
-						'url'           => $package['url'],
-						'local_package_fetch_url' => $package['fetch_url'],
-						'sha256'        => $package['sha256'],
-						'local_package' => true,
-						'provenance'    => array(
-							'schema' => 'wp-codebox/browser-plugin-provenance/v1',
-							'source' => 'runtime-plugin-remote-package',
-							'url'    => (string) ( $plugin['url'] ?? '' ),
-						),
-					)
-				);
-				continue;
-			}
-
-			if ( '' === $path ) {
-				$resolved[] = $plugin;
-				continue;
-			}
-
-			$slug = self::safe_key( (string) ( $plugin['slug'] ?? basename( $path ) ) );
-			if ( '' === $slug ) {
-				return new WP_Error( 'wp_codebox_browser_plugin_slug_missing', 'Browser plugin path specs require a slug.', array( 'status' => 400, 'field' => 'runtime.plugins', 'index' => $index ) );
-			}
-
-			if ( ! is_dir( $path ) ) {
-				return new WP_Error( 'wp_codebox_browser_plugin_path_missing', 'Browser plugin path does not exist.', array( 'status' => 400, 'field' => 'runtime.plugins', 'index' => $index, 'slug' => $slug ) );
-			}
-
-			$package = self::browser_package_component_plugin( $slug, $path );
-			if ( is_wp_error( $package ) ) {
-				return $package;
-			}
-
-			$resolved[] = array_merge(
-				$plugin,
-				array(
-					'slug'                    => $slug,
-					'url'                     => $package['url'],
-					'local_package_fetch_url' => $package['fetch_url'],
-					'sha256'                  => $package['sha256'],
-					'local_package'           => true,
-					'provenance'              => array(
-						'schema' => 'wp-codebox/browser-plugin-provenance/v1',
-						'source' => 'runtime-plugin-path',
-						'path'   => $path,
-					),
-				)
-			);
-		}
-
-		return $resolved;
-	}
-
-	/** @param array<string,mixed> $input Ability input. @param array<int,array<string,mixed>> $declared_plugins Caller/runtime plugin specs. @param array<int,mixed> $declared_components Caller/runtime component refs. @return array<int,array<string,mixed>>|WP_Error */
-	private static function browser_component_plugins( array $input, array $declared_plugins, array $declared_components ): array|WP_Error {
-		$components_required = self::browser_component_plugins_required( $input );
-		if ( ! $components_required && empty( $declared_components ) ) {
-			return array();
-		}
-
-		$paths = self::browser_component_paths( $input );
-		$registry = self::browser_runtime_component_registry();
-		$declared_slugs = array_values(
-			array_filter(
-				array_map( static fn( array $plugin ): string => self::safe_key( (string) ( $plugin['slug'] ?? '' ) ), $declared_plugins )
-			)
-		);
-		$component_slugs = self::browser_runtime_component_slugs( $declared_components, $components_required );
-
-		$plugins = array();
-		foreach ( $component_slugs as $slug ) {
-			if ( in_array( $slug, $declared_slugs, true ) ) {
-				continue;
-			}
-
-			$key  = self::browser_runtime_component_key( $slug );
-			$path = (string) ( $paths[ $key ] ?? '' );
-			if ( '' !== $path ) {
-				if ( ! is_dir( $path ) ) {
-					return new WP_Error( 'wp_codebox_browser_component_path_missing', 'Browser runtime component path does not exist.', array( 'status' => 400, 'slug' => $slug, 'path' => $path ) );
-				}
-
-				$package = self::browser_package_component_plugin( $slug, $path );
-				if ( is_wp_error( $package ) ) {
-					return $package;
-				}
-
-				$plugins[] = array(
-					'url'                     => $package['url'],
-					'slug'                    => $slug,
-					'activate'                => true,
-					'local_package'           => true,
-					'local_package_fetch_url' => $package['fetch_url'],
-					'sha256'                  => $package['sha256'],
-					'provenance'              => array(
-						'schema' => 'wp-codebox/browser-component-plugin-provenance/v1',
-						'source' => 'host-component-path',
-						'sha256' => $package['sha256'],
-					),
-				);
-				continue;
-			}
-
-			$component = is_array( $registry[ $slug ] ?? null ) ? $registry[ $slug ] : array();
-			if ( empty( $component ) ) {
-				continue;
-			}
-
-			$component['slug'] = $slug;
-			if ( 'url' === (string) ( $component['resource'] ?? 'url' ) && 'browser' !== (string) ( $component['package'] ?? '' ) ) {
-				$package = self::browser_package_remote_plugin( $slug, (string) ( $component['url'] ?? '' ), count( $plugins ), (string) ( $component['sha256'] ?? '' ) );
-				if ( is_wp_error( $package ) ) {
-					return $package;
-				}
-
-				$component['url']                     = $package['url'];
-				$component['local_package']           = true;
-				$component['local_package_fetch_url'] = $package['fetch_url'];
-				$component['sha256']                  = $package['sha256'];
-			}
-
-			$normalized = self::normalize_browser_plugins( array( $component ), 'runtime.components' );
-			if ( is_wp_error( $normalized ) ) {
-				return $normalized;
-			}
-
-			$plugins[] = $normalized[0];
-		}
-
-		return $plugins;
-	}
-
-	/** @return array<int,string> */
-	private static function browser_runtime_component_slugs( array $declared_components, bool $include_required ): array {
-		$slugs = array();
-		if ( $include_required ) {
-			$slugs = array( 'agents-api', 'data-machine', 'data-machine-code' );
-		}
-
-		foreach ( $declared_components as $component ) {
-			$slug = is_array( $component ) ? self::safe_key( (string) ( $component['slug'] ?? $component['component'] ?? $component['name'] ?? '' ) ) : self::safe_key( (string) $component );
-			if ( '' !== $slug ) {
-				$slugs[] = $slug;
-			}
-		}
-
-		return array_values( array_unique( $slugs ) );
-	}
-
-	private static function browser_runtime_component_key( string $slug ): string {
-		return str_replace( '-', '_', self::safe_key( $slug ) );
-	}
-
-	/** @return array<string,array<string,mixed>> */
-	private static function browser_runtime_component_registry(): array {
-		$registry = array(
-			'agents-api'        => array(
-				'resource'         => 'url',
-				'url'              => 'https://github.com/Automattic/agents-api/releases/latest/download/agents-api.zip',
-				'targetFolderName' => 'agents-api',
-				'activate'         => true,
-				'provenance'       => array( 'source' => 'runtime-component-registry' ),
-			),
-			'data-machine'      => array(
-				'resource'         => 'url',
-				'url'              => 'https://github.com/Extra-Chill/data-machine/releases/latest/download/data-machine.zip',
-				'targetFolderName' => 'data-machine',
-				'activate'         => true,
-				'provenance'       => array( 'source' => 'runtime-component-registry' ),
-			),
-			'data-machine-code' => array(
-				'resource'         => 'url',
-				'url'              => 'https://github.com/Extra-Chill/data-machine-code/releases/latest/download/data-machine-code.zip',
-				'targetFolderName' => 'data-machine-code',
-				'activate'         => true,
-				'provenance'       => array( 'source' => 'runtime-component-registry' ),
-			),
-		);
-
-		if ( function_exists( 'apply_filters' ) ) {
-			$registry = apply_filters( 'wp_codebox_browser_runtime_component_registry', $registry );
-		}
-
-		return is_array( $registry ) ? $registry : array();
-	}
-
-	private static function browser_component_plugins_required( array $input ): bool {
-		$provider_plugin_paths = is_array( $input['provider_plugin_paths'] ?? null ) ? $input['provider_plugin_paths'] : array();
-		$inherit               = is_array( $input['inherit'] ?? null ) ? $input['inherit'] : array();
-		$connectors            = is_array( $inherit['connectors'] ?? null ) ? $inherit['connectors'] : array();
-		$secret_env            = is_array( $input['secret_env'] ?? null ) ? $input['secret_env'] : array();
-
-		return ! empty( $input['browser_runner'] ) || ! empty( $provider_plugin_paths ) || ! empty( $connectors ) || ! empty( $secret_env );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array{agents_api:string,data_machine:string,data_machine_code:string} */
-	private static function browser_component_paths( array $input ): array {
-		$configured = self::browser_configured_component_paths();
-
-		return array(
-			'agents_api'        => self::browser_clean_path( self::browser_component_path_value( $input, 'agents_api_path', $configured['agents_api'] ?? '' ) ),
-			'data_machine'      => self::browser_clean_path( self::browser_component_path_value( $input, 'data_machine_path', $configured['data_machine'] ?? '' ) ),
-			'data_machine_code' => self::browser_clean_path( self::browser_component_path_value( $input, 'data_machine_code_path', $configured['data_machine_code'] ?? '' ) ),
-		);
-	}
-
-	private static function browser_component_path_value( array $input, string $key, string $fallback ): string {
-		$value = trim( (string) ( $input[ $key ] ?? '' ) );
-		return '' !== $value ? $value : $fallback;
-	}
-
-	/** @return array<string,mixed> */
-	private static function browser_configured_component_paths(): array {
-		$paths = array();
-		if ( function_exists( 'is_multisite' ) && is_multisite() && function_exists( 'get_site_option' ) ) {
-			$option = get_site_option( 'wp_codebox_component_paths', array() );
-		} elseif ( function_exists( 'get_option' ) ) {
-			$option = get_option( 'wp_codebox_component_paths', array() );
-		} else {
-			$option = array();
-		}
-
-		if ( is_array( $option ) ) {
-			$paths = $option;
-		}
-
-		if ( function_exists( 'apply_filters' ) ) {
-			$paths = apply_filters( 'wp_codebox_component_paths', $paths );
-		}
-
-		return is_array( $paths ) ? $paths : array();
-	}
-
-	private static function browser_clean_path( string $path ): string {
-		$path = trim( $path );
-		if ( '' === $path ) {
-			return '';
-		}
-
-		$real = realpath( $path );
-		return false !== $real ? $real : rtrim( $path, '/\\' );
-	}
-
-	/** @return array{url:string,fetch_url:string,path:string,sha256:string}|WP_Error */
-	private static function browser_package_component_plugin( string $slug, string $source_path ): array|WP_Error {
-		if ( ! class_exists( 'ZipArchive' ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_packager_missing', 'Browser runtime plugin packaging requires ZipArchive.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$upload = function_exists( 'wp_upload_dir' ) ? wp_upload_dir() : array( 'basedir' => sys_get_temp_dir(), 'baseurl' => '' );
-		if ( ! is_array( $upload ) || empty( $upload['basedir'] ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_upload_dir_missing', 'Browser runtime plugin packaging requires an upload directory.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$base_dir = rtrim( (string) $upload['basedir'], '/\\' ) . DIRECTORY_SEPARATOR . 'wp-codebox' . DIRECTORY_SEPARATOR . 'browser-runtime-plugins';
-		if ( ! is_dir( $base_dir ) && ! mkdir( $base_dir, 0777, true ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_package_dir_failed', 'Could not create browser runtime plugin package directory.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$package_id = substr( hash( 'sha256', $slug . "\n" . $source_path . "\n" . self::browser_component_source_fingerprint( $source_path ) ), 0, 16 );
-		$zip_path   = $base_dir . DIRECTORY_SEPARATOR . $slug . '-' . $package_id . '.zip';
-		if ( ! is_file( $zip_path ) ) {
-			$result = self::write_browser_plugin_zip( $slug, $source_path, $zip_path );
-			if ( is_wp_error( $result ) ) {
-				return $result;
-			}
-		}
-
-		$base_url = is_array( $upload ) && ! empty( $upload['baseurl'] ) ? rtrim( (string) $upload['baseurl'], '/' ) : '';
-		$url      = '' !== $base_url ? $base_url . '/wp-codebox/browser-runtime-plugins/' . rawurlencode( basename( $zip_path ) ) : '';
-		if ( '' === $url ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_package_url_missing', 'Browser runtime plugin package URL is missing.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$sha256 = hash_file( 'sha256', $zip_path );
-		if ( ! is_string( $sha256 ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_package_hash_failed', 'Could not hash browser runtime plugin package.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$safe_url     = self::browser_safe_local_package_url( $url );
-		$delivery_url = self::browser_plugin_delivery_url( $zip_path, $safe_url, $slug );
-		if ( is_wp_error( $delivery_url ) ) {
-			return $delivery_url;
-		}
-
-		return array(
-			'url'       => $delivery_url,
-			'fetch_url' => $safe_url,
-			'path'      => $zip_path,
-			'sha256'    => $sha256,
-		);
-	}
-
-	/** @return array{url:string,fetch_url:string,path:string,sha256:string}|WP_Error */
-	private static function browser_package_remote_plugin( string $slug, string $url, int $index, string $expected_sha256 = '' ): array|WP_Error {
-		$source = self::browser_remote_plugin_package_url( $url, $index );
-		if ( is_wp_error( $source ) ) {
-			return $source;
-		}
-
-		$expected_sha256 = strtolower( trim( $expected_sha256 ) );
-		if ( '' !== $expected_sha256 && ! preg_match( '/^[a-f0-9]{64}$/', $expected_sha256 ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_sha256_invalid', 'Browser plugin sha256 must be a 64-character hex digest.', array( 'status' => 400, 'index' => $index ) );
-		}
-
-		$upload = function_exists( 'wp_upload_dir' ) ? wp_upload_dir() : array( 'basedir' => sys_get_temp_dir(), 'baseurl' => '' );
-		if ( ! is_array( $upload ) || empty( $upload['basedir'] ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_upload_dir_missing', 'Browser runtime plugin packaging requires an upload directory.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$base_dir = rtrim( (string) $upload['basedir'], '/\\' ) . DIRECTORY_SEPARATOR . 'wp-codebox' . DIRECTORY_SEPARATOR . 'browser-runtime-plugins';
-		if ( ! is_dir( $base_dir ) && ! mkdir( $base_dir, 0777, true ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_package_dir_failed', 'Could not create browser runtime plugin package directory.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$package_id = substr( hash( 'sha256', $slug . "\n" . $source['url'] . "\n" . $expected_sha256 ), 0, 16 );
-		$zip_path   = $base_dir . DIRECTORY_SEPARATOR . $slug . '-' . $package_id . '.zip';
-		if ( ! is_file( $zip_path ) ) {
-			$downloaded = self::browser_download_remote_plugin( $source['url'], $zip_path, $slug );
-			if ( is_wp_error( $downloaded ) ) {
-				return $downloaded;
-			}
-		}
-
-		$sha256 = hash_file( 'sha256', $zip_path );
-		if ( ! is_string( $sha256 ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_package_hash_failed', 'Could not hash browser runtime plugin package.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		if ( '' !== $expected_sha256 && ! hash_equals( $expected_sha256, $sha256 ) ) {
-			@unlink( $zip_path );
-			return new WP_Error( 'wp_codebox_browser_plugin_package_hash_mismatch', 'Downloaded browser runtime plugin package does not match the expected sha256.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$base_url = is_array( $upload ) && ! empty( $upload['baseurl'] ) ? rtrim( (string) $upload['baseurl'], '/' ) : '';
-		$url      = '' !== $base_url ? $base_url . '/wp-codebox/browser-runtime-plugins/' . rawurlencode( basename( $zip_path ) ) : '';
-		if ( '' === $url ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_package_url_missing', 'Browser runtime plugin package URL is missing.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$safe_url     = self::browser_safe_local_package_url( $url );
-		$delivery_url = self::browser_plugin_delivery_url( $zip_path, $safe_url, $slug );
-		if ( is_wp_error( $delivery_url ) ) {
-			return $delivery_url;
-		}
-
-		return array(
-			'url'       => $delivery_url,
-			'fetch_url' => $safe_url,
-			'path'      => $zip_path,
-			'sha256'    => $sha256,
-		);
-	}
-
-	private static function browser_plugin_delivery_url( string $zip_path, string $public_url, string $slug ): string|WP_Error {
-		$max_bytes = (int) apply_filters( 'wp_codebox_browser_plugin_data_url_max_bytes', 16 * 1024 * 1024, $zip_path, $slug );
-		$size      = filesize( $zip_path );
-		if ( is_int( $size ) && $size > $max_bytes && ! self::browser_plugin_uses_loopback_url( $public_url ) ) {
-			return $public_url;
-		}
-
-		$contents = file_get_contents( $zip_path );
-		if ( ! is_string( $contents ) || '' === $contents ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_package_read_failed', 'Could not read browser runtime plugin package.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		return 'data:application/zip;base64,' . base64_encode( $contents );
-	}
-
-	private static function browser_safe_local_package_url( string $url ): string {
-		$parts = wp_parse_url( $url );
-		$scheme = strtolower( (string) ( $parts['scheme'] ?? '' ) );
-		if ( ! is_array( $parts ) || ! in_array( $scheme, array( 'http', 'https' ), true ) || ! self::is_loopback_host( (string) ( $parts['host'] ?? '' ) ) ) {
-			return $url;
-		}
-
-		$host = strtolower( trim( (string) $parts['host'], '[]' ) );
-		if ( 'localhost' === $host && 'http' === $scheme ) {
-			return $url;
-		}
-
-		$port     = isset( $parts['port'] ) ? ':' . (int) $parts['port'] : '';
-		$path     = (string) ( $parts['path'] ?? '' );
-		$query    = isset( $parts['query'] ) ? '?' . (string) $parts['query'] : '';
-		$fragment = isset( $parts['fragment'] ) ? '#' . (string) $parts['fragment'] : '';
-
-		return 'http://localhost' . $port . $path . $query . $fragment;
-	}
-
-	private static function browser_plugin_uses_loopback_url( string $url ): bool {
-		$parts = wp_parse_url( $url );
-		$scheme = strtolower( (string) ( $parts['scheme'] ?? '' ) );
-		return is_array( $parts ) && in_array( $scheme, array( 'http', 'https' ), true ) && self::is_loopback_host( (string) ( $parts['host'] ?? '' ) );
-	}
-
-	private static function browser_download_remote_plugin( string $url, string $zip_path, string $slug ): true|WP_Error {
-		$request = function_exists( 'wp_safe_remote_get' ) ? 'wp_safe_remote_get' : ( function_exists( 'wp_remote_get' ) ? 'wp_remote_get' : null );
-		if ( null === $request ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_http_missing', 'Browser runtime plugin remote packaging requires the WordPress HTTP API.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$response = $request(
-			$url,
-			array(
-				'timeout'     => 60,
-				'redirection' => 5,
-			)
-		);
-		if ( is_wp_error( $response ) ) {
-			return $response;
-		}
-
-		$code = function_exists( 'wp_remote_retrieve_response_code' ) ? (int) wp_remote_retrieve_response_code( $response ) : (int) ( $response['response']['code'] ?? 0 );
-		if ( $code < 200 || $code >= 300 ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_download_failed', 'Could not download browser runtime plugin package.', array( 'status' => 502, 'slug' => $slug, 'http_status' => $code ) );
-		}
-
-		$body = function_exists( 'wp_remote_retrieve_body' ) ? (string) wp_remote_retrieve_body( $response ) : (string) ( $response['body'] ?? '' );
-		if ( '' === $body ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_download_empty', 'Downloaded browser runtime plugin package is empty.', array( 'status' => 502, 'slug' => $slug ) );
-		}
-
-		if ( false === file_put_contents( $zip_path, $body ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_package_write_failed', 'Could not write browser runtime plugin package.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		return true;
-	}
-
-	private static function browser_component_source_fingerprint( string $source_path ): string {
-		$source_path = rtrim( $source_path, '/\\' );
-		$entries     = array();
-		$iterator    = self::browser_component_file_iterator( $source_path );
-		foreach ( $iterator as $file ) {
-			if ( ! $file instanceof SplFileInfo || ! $file->isFile() ) {
-				continue;
-			}
-
-			$path = $file->getPathname();
-			$relative = ltrim( substr( $path, strlen( $source_path ) ), '/\\' );
-			if ( '' === $relative ) {
-				continue;
-			}
-
-			$entries[] = str_replace( DIRECTORY_SEPARATOR, '/', $relative ) . ':' . $file->getSize() . ':' . $file->getMTime();
-		}
-
-		sort( $entries, SORT_STRING );
-
-		return hash( 'sha256', implode( "\n", $entries ) );
-	}
-
-	private static function write_browser_plugin_zip( string $slug, string $source_path, string $zip_path ): true|WP_Error {
-		$zip = new ZipArchive();
-		if ( true !== $zip->open( $zip_path, ZipArchive::CREATE | ZipArchive::OVERWRITE ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_zip_open_failed', 'Could not open browser runtime plugin package.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		$source_path = rtrim( $source_path, '/\\' );
-		$iterator    = self::browser_component_file_iterator( $source_path );
-
-		foreach ( $iterator as $file ) {
-			if ( ! $file instanceof SplFileInfo || ! $file->isFile() ) {
-				continue;
-			}
-
-			$path = $file->getPathname();
-			$relative = ltrim( substr( $path, strlen( $source_path ) ), '/\\' );
-			if ( '' === $relative ) {
-				continue;
-			}
-
-			$zip->addFile( $path, $slug . '/' . str_replace( DIRECTORY_SEPARATOR, '/', $relative ) );
-		}
-
-		if ( true !== $zip->close() ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_zip_close_failed', 'Could not close browser runtime plugin package.', array( 'status' => 500, 'slug' => $slug ) );
-		}
-
-		return true;
-	}
-
-	private static function browser_component_file_iterator( string $source_path ): RecursiveIteratorIterator {
-		return new RecursiveIteratorIterator(
-			new RecursiveCallbackFilterIterator(
-				new RecursiveDirectoryIterator( $source_path, FilesystemIterator::SKIP_DOTS ),
-				static fn( SplFileInfo $file ): bool => ! in_array( $file->getFilename(), array( '.git', '.svn', '.hg', 'node_modules' ), true )
-			)
-		);
-	}
-
-	/** @param array<int,array<string,mixed>> $plugins Browser plugin specs. @return array<int,array<string,mixed>> */
-	private static function dedupe_browser_plugins( array $plugins ): array {
-		$deduped = array();
-		$slugs   = array();
-		foreach ( $plugins as $plugin ) {
-			$slug = self::safe_key( (string) ( $plugin['slug'] ?? '' ) );
-			if ( '' !== $slug ) {
-				if ( isset( $slugs[ $slug ] ) ) {
-					continue;
-				}
-
-				$slugs[ $slug ] = true;
-			}
-
-			$deduped[] = $plugin;
-		}
-
-		return $deduped;
-	}
-
-	/** @param array<int,mixed> $plugins Plugin dependency specs. @return array<int,array<string,mixed>>|WP_Error */
-	private static function normalize_browser_plugins( array $plugins, string $field ): array|WP_Error {
-		$normalized = array();
-
-		foreach ( $plugins as $index => $plugin ) {
-			if ( ! is_array( $plugin ) ) {
-				return new WP_Error( 'wp_codebox_browser_plugin_invalid', 'Each browser plugin must be an object.', array( 'status' => 400, 'field' => $field, 'index' => $index ) );
-			}
-
-			$url      = trim( (string) ( $plugin['url'] ?? '' ) );
-			$slug     = self::safe_key( (string) ( $plugin['slug'] ?? '' ) );
-			$resource = (string) ( $plugin['resource'] ?? 'url' );
-			if ( ! in_array( $resource, array( 'url', 'git:directory' ), true ) ) {
-				return new WP_Error( 'wp_codebox_browser_plugin_resource_invalid', 'Browser plugin resource is not supported.', array( 'status' => 400, 'index' => $index, 'resource' => $resource ) );
-			}
-
-			$source = ! empty( $plugin['local_package'] ) ? self::browser_local_plugin_url( $url, $index ) : self::browser_plugin_url( $url, $index );
-			if ( is_wp_error( $source ) ) {
-				return $source;
-			}
-
-			$sha256 = strtolower( trim( (string) ( $plugin['sha256'] ?? '' ) ) );
-			if ( '' !== $sha256 && ! preg_match( '/^[a-f0-9]{64}$/', $sha256 ) ) {
-				return new WP_Error( 'wp_codebox_browser_plugin_sha256_invalid', 'Browser plugin sha256 must be a 64-character hex digest.', array( 'status' => 400, 'index' => $index ) );
-			}
-
-			$provenance = is_array( $plugin['provenance'] ?? null ) ? $plugin['provenance'] : array();
-
-			$normalized[] = array(
-				'url'              => $source['url'],
-				'slug'             => $slug,
-				'resource'         => $resource,
-				'activate'         => ! array_key_exists( 'activate', $plugin ) || (bool) $plugin['activate'],
-				'local_package'    => ! empty( $plugin['local_package'] ),
-				'local_package_fetch_url' => ! empty( $plugin['local_package'] ) ? trim( (string) ( $plugin['local_package_fetch_url'] ?? $plugin['url'] ?? '' ) ) : '',
-				'sha256'           => $sha256,
-				'ref'              => sanitize_text_field( (string) ( $plugin['ref'] ?? '' ) ),
-				'refType'          => sanitize_key( (string) ( $plugin['refType'] ?? '' ) ),
-				'path'             => 'git:directory' === $resource ? ltrim( str_replace( '\\', '/', (string) ( $plugin['path'] ?? '' ) ), '/' ) : '',
-				'targetFolderName' => sanitize_key( (string) ( $plugin['targetFolderName'] ?? '' ) ),
-				'provenance'       => array_filter(
-					array(
-						'schema' => 'wp-codebox/browser-plugin-provenance/v1',
-						'url'    => $source['url'],
-						'origin' => $source['origin'],
-						'host'   => $source['host'],
-						'source' => is_string( $provenance['source'] ?? null ) ? $provenance['source'] : '',
-						'sha256' => $sha256,
-					)
-				),
-			);
-		}
-
-		return $normalized;
-	}
-
-	/** @return array{url:string,origin:string,host:string}|WP_Error */
-	private static function browser_local_plugin_url( string $url, int $index ): array|WP_Error {
-		if ( str_starts_with( $url, 'data:application/zip;base64,' ) ) {
-			return array(
-				'url'    => $url,
-				'origin' => 'data:',
-				'host'   => 'data',
-			);
-		}
-
-		$parts = wp_parse_url( $url );
-		if ( ! is_array( $parts ) || empty( $parts['scheme'] ) || empty( $parts['host'] ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_url_invalid', 'Browser plugin URL must be absolute.', array( 'status' => 400, 'index' => $index ) );
-		}
-
-		return array(
-			'url'    => $url,
-			'origin' => self::url_origin( $parts ),
-			'host'   => strtolower( (string) $parts['host'] ),
-		);
-	}
-
-	/** @return array{url:string,origin:string,host:string}|WP_Error */
-	private static function browser_remote_plugin_package_url( string $url, int $index ): array|WP_Error {
-		$parts = wp_parse_url( $url );
-		if ( ! is_array( $parts ) || empty( $parts['scheme'] ) || empty( $parts['host'] ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_url_invalid', 'Browser plugin URL must be absolute.', array( 'status' => 400, 'index' => $index ) );
-		}
-
-		$scheme = strtolower( (string) $parts['scheme'] );
-		$host   = strtolower( (string) $parts['host'] );
-		if ( 'https' !== $scheme ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_url_insecure', 'Browser plugin URL must use https://.', array( 'status' => 400, 'index' => $index ) );
-		}
-
-		$default_hosts = array( 'downloads.wordpress.org', 'github.com', 'codeload.github.com' );
-		$allowed_hosts = array_map( 'strtolower', self::string_list( apply_filters( 'wp_codebox_browser_runtime_plugin_package_allowed_hosts', $default_hosts, $url, $index ) ) );
-		if ( ! in_array( $host, $allowed_hosts, true ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_host_not_allowed', 'Browser plugin URL host is not allowed.', array( 'status' => 400, 'index' => $index, 'host' => $host ) );
-		}
-
-		return array( 'url' => $url, 'origin' => self::url_origin( $parts ), 'host' => $host );
-	}
-
-	/** @return array{url:string,origin:string,host:string}|WP_Error */
-	private static function browser_plugin_url( string $url, int $index ): array|WP_Error {
-		$parts = wp_parse_url( $url );
-		if ( ! is_array( $parts ) || empty( $parts['scheme'] ) || empty( $parts['host'] ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_url_invalid', 'Browser plugin URL must be absolute.', array( 'status' => 400, 'index' => $index ) );
-		}
-
-		$scheme     = strtolower( (string) $parts['scheme'] );
-		$host       = strtolower( (string) $parts['host'] );
-		$allow_http = self::is_loopback_host( $host ) || (bool) apply_filters( 'wp_codebox_browser_plugin_allow_http', false, $url, $index );
-		if ( 'https' !== $scheme && ! ( $allow_http && 'http' === $scheme ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_url_insecure', 'Browser plugin URL must use https://.', array( 'status' => 400, 'index' => $index ) );
-		}
-
-		$origin        = self::url_origin( $parts );
-		$default_hosts = self::is_loopback_host( $host ) ? array( 'downloads.wordpress.org', 'github.com', 'codeload.github.com', $host ) : array( 'downloads.wordpress.org', 'github.com', 'codeload.github.com' );
-		$allowed_hosts = array_map( 'strtolower', self::string_list( apply_filters( 'wp_codebox_browser_plugin_allowed_hosts', $default_hosts, $url, $index ) ) );
-		if ( ! in_array( $host, $allowed_hosts, true ) ) {
-			return new WP_Error( 'wp_codebox_browser_plugin_host_not_allowed', 'Browser plugin URL host is not allowed.', array( 'status' => 400, 'index' => $index, 'host' => $host ) );
-		}
-
-		return array( 'url' => $url, 'origin' => $origin, 'host' => $host );
-	}
-
-	/** @return array{url:string,origin:string,host:string}|WP_Error */
-	private static function browser_theme_url( string $url, int $index ): array|WP_Error {
-		$parts = wp_parse_url( $url );
-		if ( ! is_array( $parts ) || empty( $parts['scheme'] ) || empty( $parts['host'] ) ) {
-			return new WP_Error( 'wp_codebox_browser_theme_url_invalid', 'Browser theme URL must be absolute.', array( 'status' => 400, 'index' => $index ) );
-		}
-
-		$scheme     = strtolower( (string) $parts['scheme'] );
-		$host       = strtolower( (string) $parts['host'] );
-		$allow_http = self::is_loopback_host( $host );
-		if ( 'https' !== $scheme && ! ( $allow_http && 'http' === $scheme ) ) {
-			return new WP_Error( 'wp_codebox_browser_theme_url_insecure', 'Browser theme URL must use https://.', array( 'status' => 400, 'index' => $index ) );
-		}
-
-		$origin        = self::url_origin( $parts );
-		$default_hosts = self::is_loopback_host( $host ) ? array( 'downloads.wordpress.org', $host ) : array( 'downloads.wordpress.org' );
-		$allowed_hosts = array_map( 'strtolower', self::string_list( apply_filters( 'wp_codebox_browser_theme_allowed_hosts', $default_hosts, $url, $index ) ) );
-		if ( ! in_array( $host, $allowed_hosts, true ) ) {
-			return new WP_Error( 'wp_codebox_browser_theme_host_not_allowed', 'Browser theme URL host is not allowed.', array( 'status' => 400, 'index' => $index, 'host' => $host ) );
-		}
-
-		return array( 'url' => $url, 'origin' => $origin, 'host' => $host );
-	}
-
-	/** @param array<string,string|int> $parts URL parts. */
-	private static function url_origin( array $parts ): string {
-		$scheme = strtolower( (string) ( $parts['scheme'] ?? '' ) );
-		$host   = strtolower( (string) ( $parts['host'] ?? '' ) );
-		$port   = isset( $parts['port'] ) ? ':' . (int) $parts['port'] : '';
-		return $scheme . '://' . $host . $port;
-	}
-
-	private static function is_loopback_host( string $host ): bool {
-		$host = strtolower( trim( $host, '[]' ) );
-		return 'localhost' === $host || '127.0.0.1' === $host || '::1' === $host;
-	}
-
-	/** @return string[] */
-	private static function normalized_origins( mixed $origins ): array {
-		$normalized = array();
-		foreach ( self::string_list( $origins ) as $origin ) {
-			$parts = wp_parse_url( $origin );
-			if ( is_array( $parts ) && ! empty( $parts['scheme'] ) && ! empty( $parts['host'] ) ) {
-				$normalized[] = self::url_origin( $parts );
-			}
-		}
-		return array_values( array_unique( $normalized ) );
-	}
-
-	private static function safe_key( string $value ): string {
-		if ( function_exists( 'sanitize_key' ) ) {
-			return sanitize_key( $value );
-		}
-
-		return strtolower( preg_replace( '/[^a-zA-Z0-9_-]/', '', $value ) ?? '' );
-	}
-
-	/** @param array<string,mixed> $input Ability input. @return array<string,mixed>|WP_Error */
-	private static function browser_site_blueprint_artifact( array $input ): array|WP_Error {
-		$artifact = is_array( $input['site_blueprint_artifact'] ?? null ) ? $input['site_blueprint_artifact'] : array();
-		if ( empty( $artifact ) ) {
-			return array();
-		}
-
-		$blueprint = $artifact['blueprint'] ?? null;
-		if ( ! is_array( $blueprint ) ) {
-			return new WP_Error( 'wp_codebox_site_blueprint_artifact_invalid', 'site_blueprint_artifact.blueprint must be a Playground blueprint object.', array( 'status' => 400 ) );
-		}
-
-		return array(
-			'schema'     => (string) ( $artifact['schema'] ?? 'wp-codebox/site-blueprint-artifact/v1' ),
-			'id'         => (string) ( $artifact['id'] ?? '' ),
-			'blueprint'  => $blueprint,
-			'provenance' => is_array( $artifact['provenance'] ?? null ) ? $artifact['provenance'] : array(),
-		);
-	}
-
-	/** @param array<string,mixed> $blueprint Blueprint override. @param array<string,mixed> $site_blueprint_artifact Normalized site blueprint artifact. @return array<string,mixed> */
-	private static function browser_blueprint_with_site_artifact( array $blueprint, array $site_blueprint_artifact ): array {
-		$site_blueprint = is_array( $site_blueprint_artifact['blueprint'] ?? null ) ? $site_blueprint_artifact['blueprint'] : array();
-		if ( empty( $site_blueprint ) ) {
-			return $blueprint;
-		}
-
-		$site_steps = is_array( $site_blueprint['steps'] ?? null ) ? $site_blueprint['steps'] : array();
-		$base_steps = is_array( $blueprint['steps'] ?? null ) ? $blueprint['steps'] : array();
-		$merged     = array_merge( $site_blueprint, $blueprint );
-		$merged['steps'] = array_values( array_merge( $site_steps, $base_steps ) );
-
-		if ( isset( $site_blueprint['features'] ) && isset( $blueprint['features'] ) && is_array( $site_blueprint['features'] ) && is_array( $blueprint['features'] ) ) {
-			$merged['features'] = array_merge( $site_blueprint['features'], $blueprint['features'] );
-		}
-
-		return $merged;
-	}
-
-	/** @param array<string,mixed> $blueprint Blueprint override. @param array<string,mixed> $runtime Runtime dependency specs. @return array<string,mixed> */
-	private static function browser_blueprint_with_runtime( array $blueprint, array $runtime, array $playground = array() ): array {
-		$steps = is_array( $blueprint['steps'] ?? null ) ? $blueprint['steps'] : array();
-		if ( ! self::browser_blueprint_has_login_step( $steps ) ) {
-			array_unshift(
-				$steps,
-				array(
-					'step'     => 'login',
-					'username' => 'admin',
-					'password' => 'password',
-				)
-			);
-		}
-
-		foreach ( $runtime['plugins'] as $plugin ) {
-			if ( ! empty( $plugin['local_package'] ) && self::browser_plugin_uses_loopback_url( (string) ( $plugin['url'] ?? '' ) ) ) {
-				$steps[] = array(
-					'step' => 'runPHP',
-					'code' => self::browser_plugin_install_php( $plugin ),
-				);
-				continue;
-			}
-
-			$plugin_data = array(
-				'resource' => (string) ( $plugin['resource'] ?? 'url' ),
-				'url'      => $plugin['url'],
-			);
-
-			if ( 'git:directory' === $plugin_data['resource'] ) {
-				$plugin_data['ref']     = (string) ( $plugin['ref'] ?? 'main' );
-				$plugin_data['refType'] = (string) ( $plugin['refType'] ?? 'branch' );
-				if ( '' !== (string) ( $plugin['path'] ?? '' ) ) {
-					$plugin_data['path'] = (string) $plugin['path'];
-				}
-			}
-
-			$options = array(
-				'activate' => (bool) $plugin['activate'],
-			);
-			if ( '' !== (string) ( $plugin['targetFolderName'] ?? '' ) ) {
-				$options['targetFolderName'] = (string) $plugin['targetFolderName'];
-			}
-
-			$steps[] = array(
-				'step'       => 'installPlugin',
-				'pluginData' => $plugin_data,
-				'options'    => $options,
-			);
-		}
-
-		foreach ( $runtime['mu_plugins'] as $mu_plugin ) {
-			$steps[] = array(
-				'step' => 'runPHP',
-				'code' => self::browser_mu_plugin_install_php( $mu_plugin ),
-			);
-		}
-
-		foreach ( $runtime['themes'] as $theme ) {
-			if ( ! empty( $theme['url'] ) ) {
-				$steps[] = array(
-					'step'      => 'installTheme',
-					'themeData' => array(
-						'resource' => 'url',
-						'url'      => $theme['url'],
-					),
-					'options'   => array(
-						'activate' => (bool) $theme['activate'],
-					),
-				);
-			}
-
-			if ( ! empty( $theme['files'] ) ) {
-				$steps[] = array(
-					'step' => 'runPHP',
-					'code' => self::browser_theme_files_install_php( $theme ),
-				);
-			}
-		}
-
-		foreach ( $runtime['bootstrap'] as $operation ) {
-			$steps[] = array(
-				'step' => 'runPHP',
-				'code' => self::browser_bootstrap_operation_php( $operation ),
-			);
-		}
-
-		$blueprint['steps'] = $steps;
-		if ( ! isset( $blueprint['preferredVersions'] ) ) {
-			$blueprint['preferredVersions'] = array(
-				'wp'  => (string) ( $playground['wp'] ?? 'latest' ),
-				'php' => (string) ( $playground['php'] ?? 'latest' ),
-			);
-		}
-		if ( ! isset( $blueprint['features'] ) ) {
-			$blueprint['features'] = array( 'networking' => true );
-		}
-
-		return $blueprint;
-	}
-
-	/** @param array<string,mixed> $plugin Plugin spec. */
-	private static function browser_plugin_install_php( array $plugin ): string {
-		$target_folder = sanitize_key( (string) ( $plugin['targetFolderName'] ?? $plugin['slug'] ?? '' ) );
-		if ( '' === $target_folder ) {
-			$target_folder = 'wp-codebox-runtime-plugin';
-		}
-
-		$package_url = (string) ( $plugin['url'] ?? '' );
-
-		return '<?php
-$package_url = ' . var_export( $package_url, true ) . ';
-$expected_sha256 = ' . var_export( (string) ( $plugin['sha256'] ?? '' ), true ) . ';
-$target_folder = ' . var_export( $target_folder, true ) . ';
-$activate = ' . ( ! empty( $plugin['activate'] ) ? 'true' : 'false' ) . ';
-
-$archive = str_starts_with( $package_url, "data:application/zip;base64," )
-	? base64_decode( substr( $package_url, strlen( "data:application/zip;base64," ) ), true )
-	: file_get_contents( $package_url );
-if ( ! is_string( $archive ) || "" === $archive ) {
-	throw new RuntimeException( "Could not read browser plugin package." );
-}
-if ( "" !== $expected_sha256 && ! hash_equals( $expected_sha256, hash( "sha256", $archive ) ) ) {
-	throw new RuntimeException( "Browser plugin package hash mismatch." );
-}
-
-$tmp_zip = tempnam( sys_get_temp_dir(), "wp-codebox-plugin-" );
-if ( false === $tmp_zip || false === file_put_contents( $tmp_zip, $archive ) ) {
-	throw new RuntimeException( "Could not stage browser plugin package." );
-}
-
-$zip = new ZipArchive();
-if ( true !== $zip->open( $tmp_zip ) ) {
-	@unlink( $tmp_zip );
-	throw new RuntimeException( "Could not open browser plugin package." );
-}
-
-$plugins_directory = "/wordpress/wp-content/plugins";
-if ( ! is_dir( $plugins_directory ) ) {
-	mkdir( $plugins_directory, 0777, true );
-}
-$zip->extractTo( $plugins_directory );
-$zip->close();
-@unlink( $tmp_zip );
-
-if ( $activate ) {
-	require_once "/wordpress/wp-load.php";
-	require_once ABSPATH . "wp-admin/includes/plugin.php";
-	$plugins = get_plugins( "/" . $target_folder );
-	$plugin_file = "";
-	foreach ( array_keys( $plugins ) as $file ) {
-		$plugin_file = $target_folder . "/" . $file;
-		break;
-	}
-	if ( "" === $plugin_file ) {
-		throw new RuntimeException( "Browser plugin package entry file is missing." );
-	}
-	$result = activate_plugin( $plugin_file );
-	if ( is_wp_error( $result ) ) {
-		throw new RuntimeException( $result->get_error_message() );
-	}
-}
-';
-	}
-
-	/** @param array<string,mixed> $mu_plugin Mu-plugin spec. */
-	private static function browser_mu_plugin_install_php( array $mu_plugin ): string {
-		if ( ! empty( $mu_plugin['local_package'] ) ) {
-			return self::browser_packaged_mu_plugin_install_php( $mu_plugin );
-		}
-
-		return '<?php
-$path = ' . var_export( $mu_plugin['path'], true ) . ';
-$directory = dirname( $path );
-if ( ! is_dir( $directory ) ) {
-	mkdir( $directory, 0777, true );
-}
-file_put_contents( $path, ' . var_export( $mu_plugin['content'], true ) . ' );
-';
-	}
-
-	/** @param array<string,mixed> $mu_plugin Packaged mu-plugin spec. */
-	private static function browser_packaged_mu_plugin_install_php( array $mu_plugin ): string {
-		$package_url = (string) ( $mu_plugin['url'] ?? '' );
-
-		return '<?php
-$package_url = ' . var_export( $package_url, true ) . ';
-$expected_sha256 = ' . var_export( (string) ( $mu_plugin['sha256'] ?? '' ), true ) . ';
-$target_directory = "/wordpress/wp-content/mu-plugins/" . ' . var_export( (string) $mu_plugin['targetFolderName'], true ) . ';
-$loader_path = ' . var_export( (string) $mu_plugin['path'], true ) . ';
-$entry = ' . var_export( (string) $mu_plugin['entry'], true ) . ';
-
-$archive = str_starts_with( $package_url, "data:application/zip;base64," )
-	? base64_decode( substr( $package_url, strlen( "data:application/zip;base64," ) ), true )
-	: file_get_contents( $package_url );
-if ( ! is_string( $archive ) || "" === $archive ) {
-	throw new RuntimeException( "Could not read browser mu-plugin package." );
-}
-if ( "" !== $expected_sha256 && ! hash_equals( $expected_sha256, hash( "sha256", $archive ) ) ) {
-	throw new RuntimeException( "Browser mu-plugin package hash mismatch." );
-}
-
-$tmp_zip = tempnam( sys_get_temp_dir(), "wp-codebox-mu-plugin-" );
-if ( false === $tmp_zip || false === file_put_contents( $tmp_zip, $archive ) ) {
-	throw new RuntimeException( "Could not stage browser mu-plugin package." );
-}
-
-$zip = new ZipArchive();
-if ( true !== $zip->open( $tmp_zip ) ) {
-	@unlink( $tmp_zip );
-	throw new RuntimeException( "Could not open browser mu-plugin package." );
-}
-if ( ! is_dir( $target_directory ) ) {
-	mkdir( $target_directory, 0777, true );
-}
-$zip->extractTo( dirname( $target_directory ) );
-$zip->close();
-@unlink( $tmp_zip );
-
-$entry_path = $target_directory . "/" . $entry;
-if ( ! is_readable( $entry_path ) ) {
-	throw new RuntimeException( "Browser mu-plugin package entry file is missing." );
-}
-$loader_directory = dirname( $loader_path );
-if ( ! is_dir( $loader_directory ) ) {
-	mkdir( $loader_directory, 0777, true );
-}
-file_put_contents( $loader_path, "<?php\nrequire_once " . var_export( $entry_path, true ) . ";\n" );
-';
-	}
-
-	/** @param array<string,mixed> $theme Theme spec. */
-	private static function browser_theme_files_install_php( array $theme ): string {
-		$files = array();
-		foreach ( $theme['files'] as $file ) {
-			$files[ $file['playground_path'] ] = $file['content'];
-		}
-
-		return '<?php
-$files = ' . var_export( $files, true ) . ';
-foreach ( $files as $path => $content ) {
-	$directory = dirname( $path );
-	if ( ! is_dir( $directory ) ) {
-		mkdir( $directory, 0777, true );
-	}
-	file_put_contents( $path, $content );
-}
-' . ( (bool) $theme['activate'] ? self::browser_theme_activation_php( (string) $theme['slug'] ) : '' ) . '
-';
-	}
-
-	private static function browser_theme_activation_php( string $slug ): string {
-		return self::browser_wordpress_bootstrap_php() . '
-if ( ! function_exists( \'switch_theme\' ) ) {
-	require_once ABSPATH . WPINC . \'/theme.php\';
-}
-switch_theme( ' . var_export( $slug, true ) . ' );';
-	}
-
-	private static function browser_wordpress_bootstrap_php(): string {
-		return 'if ( ! defined( \'ABSPATH\' ) ) {
-	require_once \'/wordpress/wp-load.php\';
-}';
-	}
-
-	/** @param array<string,mixed> $operation Bootstrap operation spec. */
-	private static function browser_bootstrap_operation_php( array $operation ): string {
-		$args = is_array( $operation['args'] ?? null ) ? $operation['args'] : array();
-		switch ( $operation['operation'] ) {
-			case 'set_option':
-				return '<?php
-' . self::browser_wordpress_bootstrap_php() . '
-update_option( ' . var_export( (string) ( $args['name'] ?? $args['option'] ?? '' ), true ) . ', ' . var_export( $args['value'] ?? '', true ) . ' );
-';
-			case 'activate_plugin':
-				return '<?php
-' . self::browser_wordpress_bootstrap_php() . '
-activate_plugin( ' . var_export( (string) ( $args['plugin'] ?? '' ), true ) . ' );
-';
-			case 'activate_theme':
-				return '<?php
-' . self::browser_theme_activation_php( (string) ( $args['slug'] ?? $args['theme'] ?? '' ) ) . '
-';
-			case 'flush_rewrite_rules':
-				return '<?php
-' . self::browser_wordpress_bootstrap_php() . '
-flush_rewrite_rules();
-';
-		}
-
-		return '<?php';
-	}
-
-	/** @param array<string,mixed> $runtime Runtime dependency specs. @return array<string,mixed> */
-	private static function browser_runtime_readiness_metadata( array $runtime ): array {
-		return array(
-			'schema'    => 'wp-codebox/browser-runtime-readiness/v1',
-			'compiled'  => true,
-			'summary'   => $runtime['summary'] ?? array(),
-			'plugins'   => array_map( static fn( array $plugin ): array => array( 'slug' => $plugin['slug'] ?? '', 'activate' => (bool) ( $plugin['activate'] ?? true ), 'readiness' => 'compiled' ), $runtime['plugins'] ?? array() ),
-			'mu_plugins' => array_map( static fn( array $mu_plugin ): array => array( 'slug' => $mu_plugin['slug'] ?? '', 'file' => $mu_plugin['file'] ?? '', 'readiness' => $mu_plugin['readiness'] ?? 'compiled' ), $runtime['mu_plugins'] ?? array() ),
-			'themes'    => array_map( static fn( array $theme ): array => array( 'slug' => $theme['slug'] ?? '', 'activate' => (bool) ( $theme['activate'] ?? true ), 'readiness' => $theme['readiness'] ?? 'compiled' ), $runtime['themes'] ?? array() ),
-			'bootstrap' => array_map( static fn( array $operation ): array => array( 'operation' => $operation['operation'] ?? '', 'readiness' => $operation['readiness'] ?? 'compiled' ), $runtime['bootstrap'] ?? array() ),
-		);
-	}
-
-	/** @param array<int,mixed> $steps Blueprint steps. */
-	private static function browser_blueprint_has_login_step( array $steps ): bool {
-		foreach ( $steps as $step ) {
-			if ( is_array( $step ) && 'login' === (string) ( $step['step'] ?? '' ) ) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	/** @param array<string,mixed> $task_input Normalized task input. @param array<string,mixed> $runner Runner overrides. @param array<string,mixed> $task_payload Browser task payload. @return array<string,mixed>|WP_Error */
-	private static function browser_agent_recipe( array $task_input, string $session_id, array $runner, array $blueprint, array $playground, array $task_payload ): array|WP_Error {
-		$task_path   = (string) ( $runner['task_path'] ?? '/tmp/wp-codebox-agent-task.json' );
-		$result_path = (string) ( $runner['result_path'] ?? '/tmp/wp-codebox-agent-result.json' );
-		$invocation  = self::browser_runner_invocation( $runner );
-		if ( is_wp_error( $invocation ) ) {
-			return $invocation;
-		}
-		$captures = self::browser_runner_capture_paths( $runner );
-		if ( is_wp_error( $captures ) ) {
-			return $captures;
-		}
-
-		foreach ( array( 'task_path' => $task_path, 'result_path' => $result_path ) as $field => $path ) {
-			if ( '' === $path || str_contains( $path, '..' ) || ! str_starts_with( $path, '/' ) || ! preg_match( '#^[A-Za-z0-9_./-]+$#', $path ) ) {
-				return new WP_Error( 'wp_codebox_browser_runner_path_invalid', $field . ' must be a safe absolute Playground path.', array( 'status' => 400 ) );
-			}
-		}
-
-		return array(
-			'schema'   => 'wp-codebox/workspace-recipe/v1',
-			'runtime'  => array(
-				'backend'   => 'wordpress-playground',
-				'name'      => 'browser-playground',
-				'wp'        => (string) ( $playground['wp'] ?? 'latest' ),
-				'blueprint' => self::browser_playground_blueprint( $blueprint, $playground ),
-			),
-			'inputs'   => array(
-				'stagedFiles' => array(
-					array(
-						'source' => 'task-payload',
-						'target' => $task_path,
-					),
-				),
-				'agent_bundles' => is_array( $task_payload['agent_bundles'] ?? null ) ? $task_payload['agent_bundles'] : array(),
-			),
-			'workflow' => array(
-				'steps' => array(
-					array(
-						'command' => 'wordpress.run-php',
-						'args'    => array(
-							'code=' . self::browser_agent_runner_php( $task_input, $session_id, $task_path, $result_path, $invocation, $captures ),
-						),
-					),
-				),
-			),
-			'artifacts' => array(
-				'directory' => self::browser_artifact_base_path( $playground ),
-			),
-			'browser'  => array(
-				'execution'  => 'php-wasm',
-				'task_path'  => $task_path,
-				'result_path' => $result_path,
-				'task_payload' => $task_payload,
-				'invocation' => self::browser_runner_invocation_metadata( $invocation ),
-				'captures'   => $captures,
-			),
-		);
-	}
-
-	/** @param array<string,mixed> $recipe Browser recipe. @return array<string,mixed> */
-	private static function browser_materialization_contract( array $recipe ): array {
-		$browser = is_array( $recipe['browser'] ?? null ) ? $recipe['browser'] : array();
-		return array(
-			'schema'        => 'wp-codebox/browser-materialization/v1',
-			'status'        => 'pending',
-			'execution'     => (string) ( $browser['execution'] ?? 'php-wasm' ),
-			'result_path'   => (string) ( $browser['result_path'] ?? '' ),
-			'invocation'    => is_array( $browser['invocation'] ?? null ) ? $browser['invocation'] : array(),
-			'captures'      => is_array( $browser['captures'] ?? null ) ? $browser['captures'] : array(),
-			'diagnostics'   => array(
-				'capture_count' => count( is_array( $browser['captures'] ?? null ) ? $browser['captures'] : array() ),
-			),
-			'errors'        => array(),
-			'provenance'    => array(
-				'generated_by' => 'wp-codebox/browser-runner',
-				'task_path'    => (string) ( $browser['task_path'] ?? '' ),
-				'result_path'  => (string) ( $browser['result_path'] ?? '' ),
-			),
-			'error_schema'  => 'wp-codebox/browser-materialization-error/v1',
-		);
-	}
-
-	/** @param array<string,mixed> $runner Runner overrides. @return array<string,mixed>|WP_Error */
-	private static function browser_runner_invocation( array $runner ): array|WP_Error {
-		$invocation = is_array( $runner['invocation'] ?? null ) ? $runner['invocation'] : array();
-		$type       = self::safe_key( (string) ( $invocation['type'] ?? 'ability' ) );
-		if ( ! in_array( $type, array( 'ability', 'task' ), true ) ) {
-			return new WP_Error( 'wp_codebox_browser_invocation_type_invalid', 'Browser runner invocation type must be ability or task.', array( 'status' => 400 ) );
-		}
-
-		$name = trim( (string) ( $invocation['name'] ?? '' ) );
-		$hook = trim( (string) ( $invocation['hook'] ?? $name ) );
-		if ( 'ability' === $type ) {
-			$name = '' !== $name ? $name : 'agents/chat';
-			if ( ! preg_match( '#^[A-Za-z0-9_.-]+/[A-Za-z0-9_.-]+$#', $name ) ) {
-				return new WP_Error( 'wp_codebox_browser_invocation_name_invalid', 'Browser runner ability names must use namespace/name form.', array( 'status' => 400 ) );
-			}
-		} elseif ( '' === $hook || ! preg_match( '#^[A-Za-z0-9_.:-]+$#', $hook ) ) {
-			return new WP_Error( 'wp_codebox_browser_invocation_hook_invalid', 'Browser runner task hooks must be safe WordPress hook names.', array( 'status' => 400 ) );
-		}
-
-		$permission_filter = trim( (string) ( $invocation['permission_filter'] ?? ( 'agents/chat' === $name ? 'agents_chat_permission' : '' ) ) );
-		if ( '' !== $permission_filter && ! preg_match( '#^[A-Za-z0-9_.:-]+$#', $permission_filter ) ) {
-			return new WP_Error( 'wp_codebox_browser_invocation_permission_filter_invalid', 'Browser runner permission filters must be safe WordPress hook names.', array( 'status' => 400 ) );
-		}
-
-		return array(
-			'type'              => $type,
-			'name'              => $name,
-			'hook'              => $hook,
-			'input'             => is_array( $invocation['input'] ?? null ) ? $invocation['input'] : array(),
-			'permission_filter' => $permission_filter,
-		);
-	}
-
-	/** @param array<string,mixed> $runner Runner overrides. @return array<int,array<string,mixed>>|WP_Error */
-	private static function browser_runner_capture_paths( array $runner ): array|WP_Error {
-		$captures = is_array( $runner['capture_paths'] ?? null ) ? $runner['capture_paths'] : array();
-		$normalized = array();
-		foreach ( $captures as $index => $capture ) {
-			if ( ! is_array( $capture ) ) {
-				return new WP_Error( 'wp_codebox_browser_capture_invalid', 'Each browser runner capture path must be an object.', array( 'status' => 400, 'index' => $index ) );
-			}
-			$path = (string) ( $capture['path'] ?? '' );
-			if ( '' === $path || str_contains( $path, '..' ) || ! str_starts_with( $path, '/' ) || ! preg_match( '#^[A-Za-z0-9_./-]+$#', $path ) ) {
-				return new WP_Error( 'wp_codebox_browser_capture_path_invalid', 'Browser runner capture paths must be safe absolute Playground paths.', array( 'status' => 400, 'index' => $index ) );
-			}
-			$max_bytes = (int) ( $capture['max_bytes'] ?? self::BROWSER_CAPTURE_MAX_BYTES );
-			if ( $max_bytes < 0 || $max_bytes > self::BROWSER_ARTIFACT_MAX_BYTES ) {
-				return new WP_Error( 'wp_codebox_browser_capture_max_bytes_invalid', 'Browser runner capture max_bytes must be between 0 and the browser artifact byte limit.', array( 'status' => 400, 'index' => $index ) );
-			}
-
-			$normalized[] = array_filter(
-				array(
-					'path'      => $path,
-					'name'      => sanitize_key( (string) ( $capture['name'] ?? '' ) ),
-					'kind'      => sanitize_key( (string) ( $capture['kind'] ?? 'report' ) ),
-					'mime_type' => sanitize_text_field( (string) ( $capture['mime_type'] ?? '' ) ),
-					'max_bytes' => $max_bytes,
-				),
-				static fn( mixed $value ): bool => '' !== $value
-			);
-		}
-
-		return $normalized;
-	}
-
-	/** @param array<string,mixed> $invocation Normalized invocation. @return array<string,string> */
-	private static function browser_runner_invocation_metadata( array $invocation ): array {
-		return array_filter(
-			array(
-				'type' => (string) $invocation['type'],
-				'name' => (string) $invocation['name'],
-				'hook' => (string) $invocation['hook'],
-			),
-			static fn( string $value ): bool => '' !== $value
-		);
-	}
-
-	private static function browser_agent_runner_php( array $task_input, string $session_id, string $task_path, string $result_path, array $invocation, array $captures ): string {
-		$default_payload = array(
-			'agent'      => 'wp-codebox-sandbox',
-			'message'    => (string) $task_input['goal'],
-			'session_id' => $session_id,
-			'task_input' => $task_input,
-			'artifacts'  => array(),
-		);
-		$default_invocation = $invocation;
-		$default_captures   = $captures;
-
-		return '<?php
-require_once \'/wordpress/wp-load.php\';
-
-$task_path = ' . var_export( $task_path, true ) . ';
-$result_path = ' . var_export( $result_path, true ) . ';
-$payload = ' . var_export( $default_payload, true ) . ';
-$invocation = ' . var_export( $default_invocation, true ) . ';
-$capture_paths = ' . var_export( $default_captures, true ) . ';
-$started_at = gmdate( \'c\' );
-
-function wp_codebox_browser_normalize_error( $error ) {
-	if ( is_wp_error( $error ) ) {
-		return array(
-			\'schema\' => \'wp-codebox/browser-materialization-error/v1\',
-			\'code\' => $error->get_error_code(),
-			\'message\' => $error->get_error_message(),
-			\'data\' => $error->get_error_data(),
-		);
-	}
-
-	return array(
-		\'schema\' => \'wp-codebox/browser-materialization-error/v1\',
-		\'code\' => \'wp_codebox_browser_runner_exception\',
-		\'message\' => $error instanceof Throwable ? $error->getMessage() : (string) $error,
-	);
-}
-
-function wp_codebox_browser_capture_file( array $capture ) {
-	$path = (string) ( $capture[\'path\'] ?? \'\' );
-	$record = array(
-		\'schema\' => \'wp-codebox/browser-capture/v1\',
-		\'path\' => $path,
-		\'name\' => (string) ( $capture[\'name\'] ?? \'\' ),
-		\'kind\' => (string) ( $capture[\'kind\'] ?? \'report\' ),
-		\'mime_type\' => (string) ( $capture[\'mime_type\'] ?? \'\' ),
-		\'exists\' => is_readable( $path ),
-	);
-	if ( ! $record[\'exists\'] ) {
-		return $record;
-	}
-	$contents = file_get_contents( $path );
-	if ( ! is_string( $contents ) ) {
-		$record[\'error\'] = array( \'code\' => \'wp_codebox_browser_capture_read_failed\', \'message\' => \'Could not read captured browser materialization file.\' );
-		return $record;
-	}
-	$size = strlen( $contents );
-	$max_bytes = isset( $capture[\'max_bytes\'] ) ? (int) $capture[\'max_bytes\'] : ' . self::BROWSER_CAPTURE_MAX_BYTES . ';
-	$record[\'size\'] = $size;
-	$record[\'sha256\'] = hash( \'sha256\', $contents );
-	$record[\'truncated\'] = $size > $max_bytes;
-	if ( $max_bytes > 0 ) {
-		$body = $record[\'truncated\'] ? substr( $contents, 0, $max_bytes ) : $contents;
-		$json = json_decode( $body, true );
-		if ( JSON_ERROR_NONE === json_last_error() ) {
-			$record[\'json\'] = $json;
-		} elseif ( preg_match( \'#^[\\x09\\x0A\\x0D\\x20-\\x7E]*$#\', $body ) ) {
-			$record[\'content\'] = $body;
-		} else {
-			$record[\'content_base64\'] = base64_encode( $body );
-			$record[\'encoding\'] = \'base64\';
-		}
-	}
-	return array_filter( $record, static fn( $value ) => \'\' !== $value );
-}
-
-function wp_codebox_browser_safe_artifact_path( $path, $root ) {
-	$path = str_replace( chr( 92 ), "/", (string) $path );
-	$path = ltrim( preg_replace( "#/+#", "/", $path ), "/" );
-	$root = str_replace( chr( 92 ), "/", (string) $root );
-	$root = rtrim( ltrim( preg_replace( "#/+#", "/", $root ), "/" ), "/" ) . "/";
-	$parts = array_filter( explode( "/", $path ), static fn( $part ) => \'\' !== $part );
-	if ( empty( $parts ) || in_array( \'..\', $parts, true ) || in_array( \'.\', $parts, true ) ) {
-		return \'\';
-	}
-	$normalized = implode( "/", $parts );
-	return str_starts_with( $normalized, $root ) ? $normalized : \'\';
-}
-
-function wp_codebox_browser_safe_playground_read_path( $path ) {
-	$path = str_replace( chr( 92 ), "/", (string) $path );
-	$root = defined( \'ABSPATH\' ) ? wp_normalize_path( ABSPATH ) : \'/wordpress/\';
-	$root = rtrim( str_replace( chr( 92 ), "/", $root ), "/" ) . "/";
-	if ( \'\' === $path || str_contains( $path, \'..\' ) || ! str_starts_with( $path, $root ) || ! is_readable( $path ) ) {
-		return \'\';
-	}
-
-	return $path;
-}
-
-function wp_codebox_browser_capture_artifact_bundle( array $payload ) {
-	$contract = is_array( $payload[\'artifacts\'] ?? null ) ? $payload[\'artifacts\'] : array();
-	$schema = trim( (string) ( $contract[\'schema\'] ?? \'\' ) );
-	if ( \'\' === $schema ) {
-		return array();
-	}
-
-	$root = (string) ( $contract[\'root\'] ?? \'\' );
-	$root = rtrim( ltrim( str_replace( chr( 92 ), "/", $root ), "/" ), "/" ) . "/";
-	if ( \'/\' === $root ) {
-		return array();
-	}
-	$entrypoint = wp_codebox_browser_safe_artifact_path( (string) ( $contract[\'entrypoint\'] ?? $root . \'index.html\' ), $root );
-	if ( \'\' === $entrypoint ) {
-		return array();
-	}
-
-	$files = array();
-	foreach ( is_array( $contract[\'files\'] ?? null ) ? $contract[\'files\'] : array() as $file ) {
-		if ( ! is_array( $file ) ) {
-			continue;
-		}
-		$path = wp_codebox_browser_safe_artifact_path( $file[\'path\'] ?? \'\', $root );
-		$playground_path = wp_codebox_browser_safe_playground_read_path( $file[\'playground_path\'] ?? \'\' );
-		if ( \'\' === $path || \'\' === $playground_path ) {
-			continue;
-		}
-		$contents = file_get_contents( $playground_path );
-		if ( ! is_string( $contents ) ) {
-			continue;
-		}
-		$is_text = 1 === preg_match( \'#^[\\x09\\x0A\\x0D\\x20-\\x7E]*$#\', $contents );
-		$record = array_diff_key( $file, array( \'playground_path\' => true, \'content\' => true, \'content_base64\' => true, \'encoding\' => true, \'sha256\' => true ) );
-		$record[\'path\'] = $path;
-		$record[\'kind\'] = (string) ( $file[\'kind\'] ?? \'\' );
-		$record[\'mime_type\'] = (string) ( $file[\'mime_type\'] ?? \'\' );
-		$record[\'url_path\'] = (string) ( $file[\'url_path\'] ?? \'\' );
-		$record[\'sha256\'] = hash( \'sha256\', $contents );
-		if ( $is_text ) {
-			$record[\'content\'] = $contents;
-			$record[\'encoding\'] = \'utf-8\';
-		} else {
-			$record[\'content_base64\'] = base64_encode( $contents );
-			$record[\'encoding\'] = \'base64\';
-		}
-		$files[] = array_filter( $record, static fn( $value ) => \'\' !== $value );
-	}
-
-	$entrypoint_captured = in_array( $entrypoint, array_column( $files, \'path\' ), true );
-	if ( empty( $files ) || ! $entrypoint_captured ) {
-		return array();
-	}
-
-	return array_merge(
-		array_diff_key( $contract, array( \'files\' => true ) ),
-		array(
-			\'schema\' => $schema,
-			\'root\' => $root,
-			\'entrypoint\' => $entrypoint,
-			\'files\' => $files,
-		)
-	);
-}
-
-function wp_codebox_browser_import_agent_bundles( array $bundle_specs ): array {
-	if ( empty( $bundle_specs ) ) {
-		return array();
-	}
-	$ability = function_exists( \'wp_get_ability\' ) ? wp_get_ability( \'datamachine/import-agent\' ) : null;
-	if ( ! $ability instanceof WP_Ability ) {
-		return array( array(
-			\'success\' => false,
-			\'error\' => array(
-				\'code\' => \'datamachine_import_agent_unavailable\',
-				\'message\' => \'The canonical datamachine/import-agent ability is not available inside the Playground site.\',
-			),
-		) );
-	}
-
-	$imports = array();
-	foreach ( $bundle_specs as $index => $spec ) {
-		if ( ! is_array( $spec ) ) {
-			$imports[] = array( \'success\' => false, \'index\' => $index, \'error\' => array( \'code\' => \'agent_bundle_spec_invalid\', \'message\' => \'Agent bundle spec must be an object.\' ) );
-			continue;
-		}
-		$source = isset( $spec[\'source\'] ) ? trim( (string) $spec[\'source\'] ) : \'\';
-		$temp_source = \'\';
-		if ( \'\' === $source && isset( $spec[\'bundle\'] ) && is_array( $spec[\'bundle\'] ) ) {
-			$temp_source = tempnam( sys_get_temp_dir(), \'wp-codebox-agent-bundle-\' );
-			if ( false === $temp_source ) {
-				$imports[] = array( \'success\' => false, \'index\' => $index, \'error\' => array( \'code\' => \'agent_bundle_temp_failed\', \'message\' => \'Could not create a temporary agent bundle JSON file.\' ) );
-				continue;
-			}
-			$json_source = wp_json_encode( $spec[\'bundle\'], JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE );
-			if ( ! is_string( $json_source ) || false === file_put_contents( $temp_source, $json_source ) ) {
-				@unlink( $temp_source );
-				$imports[] = array( \'success\' => false, \'index\' => $index, \'error\' => array( \'code\' => \'agent_bundle_write_failed\', \'message\' => \'Could not stage inline agent bundle JSON.\' ) );
-				continue;
-			}
-			$source = $temp_source;
-		}
-		if ( \'\' === $source ) {
-			$imports[] = array( \'success\' => false, \'index\' => $index, \'error\' => array( \'code\' => \'agent_bundle_source_missing\', \'message\' => \'Agent bundle spec requires source or bundle.\' ) );
-			continue;
-		}
-
-		$input = array( \'source\' => $source, \'on_conflict\' => (string) ( $spec[\'on_conflict\'] ?? \'upgrade\' ) );
-		foreach ( array( \'slug\', \'token_env\' ) as $field ) {
-			if ( isset( $spec[ $field ] ) && \'\' !== trim( (string) $spec[ $field ] ) ) {
-				$input[ $field ] = trim( (string) $spec[ $field ] );
-			}
-		}
-		$input[\'owner_id\'] = isset( $spec[\'owner_id\'] ) && (int) $spec[\'owner_id\'] > 0 ? (int) $spec[\'owner_id\'] : ( get_current_user_id() ?: 1 );
-		$result = $ability->execute( $input );
-		if ( \'\' !== $temp_source ) {
-			@unlink( $temp_source );
-		}
-		$imports[] = is_wp_error( $result )
-			? array( \'success\' => false, \'index\' => $index, \'source\' => isset( $spec[\'source\'] ) ? $source : \'inline\', \'error\' => array( \'code\' => $result->get_error_code(), \'message\' => $result->get_error_message(), \'data\' => $result->get_error_data() ) )
-			: array_merge( array( \'index\' => $index, \'source\' => isset( $spec[\'source\'] ) ? $source : \'inline\' ), is_array( $result ) ? $result : array( \'result\' => $result ) );
-	}
-
-	return $imports;
-}
-
-$wp_codebox_playground_root = defined( \'ABSPATH\' ) ? wp_normalize_path( ABSPATH ) : \'\';
-$wp_codebox_is_playground = \'/wordpress/\' === $wp_codebox_playground_root && ( \'Emscripten\' === PHP_OS_FAMILY || ( defined( \'WP_CODEBOX_BROWSER_PLAYGROUND_RUNNER\' ) && WP_CODEBOX_BROWSER_PLAYGROUND_RUNNER ) );
-
-if ( function_exists( \'wp_register_ability\' ) ) {
-	if ( ! did_action( \'wp_abilities_api_categories_init\' ) ) {
-		do_action( \'wp_abilities_api_categories_init\' );
-	}
-	if ( ! did_action( \'wp_abilities_api_init\' ) ) {
-		do_action( \'wp_abilities_api_init\' );
-	}
-}
-
-if ( is_readable( $task_path ) ) {
-	$raw_payload = json_decode( (string) file_get_contents( $task_path ), true );
-	if ( is_array( $raw_payload ) ) {
-		$payload = array_replace_recursive( $payload, $raw_payload );
-	}
-}
-
-$agent = sanitize_key( (string) ( $payload[\'agent\'] ?? \'wp-codebox-sandbox\' ) );
-$message = (string) ( $payload[\'message\'] ?? ( $payload[\'task_input\'][\'goal\'] ?? \'\' ) );
-$session_id = (string) ( $payload[\'session_id\'] ?? ' . var_export( $session_id, true ) . ' );
-$agent_bundle_imports = array();
-$base_input = array(
-	\'agent\' => $agent,
-	\'message\' => $message,
-	\'session_id\' => $session_id,
-	\'session_owner\' => array(
-		\'type\' => \'browser-playground\',
-		\'key\' => $session_id,
-		\'label\' => \'WP Codebox Browser Playground\',
-	),
-	\'client_context\' => array(
-		\'source\' => \'peer-agent\',
-		\'client_name\' => \'wp-codebox-browser-runner\',
-		\'peer_agent_call\' => true,
-		\'task_input\' => $payload[\'task_input\'] ?? array(),
-	),
-);
-$input = array_replace_recursive( $base_input, is_array( $invocation[\'input\'] ?? null ) ? $invocation[\'input\'] : array() );
-$invocation_type = (string) ( $invocation[\'type\'] ?? \'ability\' );
-$permission_filter = (string) ( $invocation[\'permission_filter\'] ?? \'\' );
-
-if ( ! $wp_codebox_is_playground ) {
-	$response = new WP_Error(
-		\'wp_codebox_browser_runner_not_playground\',
-		\'The browser agent runner permission bypass is only allowed inside the disposable WordPress Playground sandbox.\',
-		array(
-			\'execution_scope\' => \'disposable-playground\',
-			\'permission_model\' => \'sandbox-bypass\',
-			\'detected_root\' => $wp_codebox_playground_root,
-			\'detected_php_os_family\' => PHP_OS_FAMILY,
-		)
-	);
-} else {
-	$agent_bundle_imports = wp_codebox_browser_import_agent_bundles( is_array( $payload[\'agent_bundles\'] ?? null ) ? $payload[\'agent_bundles\'] : array() );
-	if ( \'\' !== $permission_filter ) {
-		add_filter( $permission_filter, \'__return_true\', 999 );
-	}
-
-	try {
-		$failed_imports = array_filter( $agent_bundle_imports, static fn( $import ) => is_array( $import ) && empty( $import[\'success\'] ) );
-		if ( ! empty( $failed_imports ) ) {
-			$response = new WP_Error( \'wp_codebox_agent_bundle_import_failed\', \'One or more Data Machine agent bundles failed to import before sandbox invocation.\', array( \'agent_bundle_imports\' => array_values( $failed_imports ) ) );
-		} elseif ( \'task\' === $invocation_type ) {
-			$hook = (string) ( $invocation[\'hook\'] ?? $invocation[\'name\'] ?? \'\' );
-			if ( \'\' === $hook || ! has_filter( $hook ) ) {
-				$response = new WP_Error( \'wp_codebox_browser_task_unavailable\', \'The requested sandbox task hook is not registered inside the Playground site.\', array( \'hook\' => $hook ) );
-			} else {
-				$response = apply_filters( $hook, null, $input, $payload );
-			}
-		} else {
-			$ability_name = (string) ( $invocation[\'name\'] ?? \'agents/chat\' );
-			$ability = function_exists( \'wp_get_ability\' ) ? wp_get_ability( $ability_name ) : null;
-			if ( ! $ability instanceof WP_Ability ) {
-				$response = new WP_Error( \'wp_codebox_browser_ability_unavailable\', \'The requested ability is not available inside the Playground site.\', array( \'ability\' => $ability_name ) );
-			} else {
-				$response = $ability->execute( $input );
-			}
-		}
-	} catch ( Throwable $exception ) {
-		$response = $exception;
-	} finally {
-		if ( \'\' !== $permission_filter ) {
-			remove_filter( $permission_filter, \'__return_true\', 999 );
-		}
-	}
-}
-
-$captures = array();
-foreach ( $capture_paths as $capture ) {
-	if ( is_array( $capture ) ) {
-		$captures[] = wp_codebox_browser_capture_file( $capture );
-	}
-}
-$artifact_bundle = wp_codebox_browser_capture_artifact_bundle( $payload );
-
-$invocation_metadata = array_filter(
-	array(
-		\'type\' => $invocation_type,
-		\'name\' => (string) ( $invocation[\'name\'] ?? \'\' ),
-		\'hook\' => (string) ( $invocation[\'hook\'] ?? \'\' ),
-	),
-	static fn( $value ) => \'\' !== $value
-);
-$diagnostics = array(
-	\'capture_count\' => count( $captures ),
-	\'captured_paths\' => array_values( array_map( static fn( $capture ) => (string) ( $capture[\'path\'] ?? \'\' ), $captures ) ),
-	\'agent_bundle_imports\' => $agent_bundle_imports,
-);
-$provenance = array(
-	\'generated_by\' => \'wp-codebox/browser-runner\',
-	\'session_id\' => $session_id,
-	\'task_path\' => $task_path,
-	\'result_path\' => $result_path,
-	\'started_at\' => $started_at,
-	\'completed_at\' => gmdate( \'c\' ),
-);
-
-if ( is_wp_error( $response ) || $response instanceof Throwable ) {
-	$error = array(
-		\'code\' => wp_codebox_browser_normalize_error( $response )[\'code\'],
-		\'message\' => wp_codebox_browser_normalize_error( $response )[\'message\'],
-		\'data\' => wp_codebox_browser_normalize_error( $response )[\'data\'] ?? null,
-	);
-	$result = array(
-			\'success\' => false,
-			\'schema\' => \'wp-codebox/browser-materialization/v1\',
-			\'status\' => \'error\',
-			\'session_id\' => $session_id,
-			\'execution_scope\' => \'disposable-playground\',
-			\'permission_model\' => \'sandbox-bypass\',
-			\'invocation\' => $invocation_metadata,
-			\'captures\' => $captures,
-			\'diagnostics\' => $diagnostics,
-			\'error\' => $error,
-			\'errors\' => array( $error ),
-			\'provenance\' => $provenance,
-	);
-} else {
-	$result = array(
-			\'success\' => true,
-			\'schema\' => \'wp-codebox/browser-materialization/v1\',
-			\'status\' => \'completed\',
-			\'session_id\' => $session_id,
-			\'execution_scope\' => \'disposable-playground\',
-			\'permission_model\' => \'sandbox-bypass\',
-			\'invocation\' => $invocation_metadata,
-			\'task_input\' => $payload[\'task_input\'] ?? array(),
-			\'response\' => $response,
-			\'captures\' => $captures,
-			\'diagnostics\' => $diagnostics,
-			\'errors\' => array(),
-			\'provenance\' => $provenance,
-			\'artifacts\' => $payload[\'artifacts\'] ?? array(),
-	);
-	if ( ! empty( $artifact_bundle ) ) {
-		$result[\'artifact_bundle\'] = $artifact_bundle;
-		if ( is_array( $result[\'response\'] ?? null ) && empty( $result[\'response\'][\'artifact_bundle\'] ) ) {
-			$result[\'response\'][\'artifact_bundle\'] = $artifact_bundle;
-		}
-	}
-}
-
-file_put_contents( $result_path, wp_json_encode( $result ) );
-echo wp_json_encode( $result );
-';
-	}
-
-	/** @param array<string,mixed> $playground Playground config. */
-	private static function browser_artifact_base_path( array $playground ): string {
-		return self::normalize_absolute_browser_path( (string) ( $playground['artifact_base_path'] ?? '/wordpress/wp-content/uploads/wp-codebox/artifacts' ) );
-	}
-
-	/** @param array<string,mixed> $playground Playground config. */
-	private static function browser_artifact_base_url( array $playground ): string {
-		return self::normalize_absolute_browser_path( (string) ( $playground['artifact_base_url'] ?? '/wp-content/uploads/wp-codebox/artifacts' ) );
-	}
-
-	/** @param array<int,array<string,string>> $artifacts Artifact files. @param array<string,mixed> $playground Playground config. */
-	private static function browser_preview_url( array $artifacts, array $playground ): string {
-		$preview_url = trim( (string) ( $playground['preview_url'] ?? '' ) );
-		if ( '' !== $preview_url ) {
-			return self::normalize_absolute_browser_path( $preview_url );
-		}
-
-		$first_file = $artifacts[0]['url_path'] ?? '';
-		return '' !== $first_file ? $first_file : '/';
-	}
-
-	private static function normalize_absolute_browser_path( string $path ): string {
-		$path = '/' . ltrim( trim( $path ), '/' );
-		$path = rtrim( $path, '/' );
-		return '' === $path ? '/' : $path;
-	}
-
-	private static function join_browser_path( string $base, string $path ): string {
-		return rtrim( $base, '/' ) . '/' . ltrim( $path, '/' );
-	}
-
-	/** @param array<string,mixed> $blueprint Blueprint override. @param array<string,mixed> $playground Playground config. @return array<string,mixed> */
-	private static function browser_playground_blueprint( array $blueprint, array $playground ): array {
-		if ( ! empty( $blueprint ) ) {
-			return $blueprint;
-		}
-
-		return array(
-			'preferredVersions' => array(
-				'wp'  => (string) ( $playground['wp'] ?? 'latest' ),
-				'php' => (string) ( $playground['php'] ?? 'latest' ),
-			),
-			'features'          => array(
-				'networking' => true,
-			),
-			'steps'             => array(),
-		);
-	}
-
-	private static function generate_id(): string {
-		if ( function_exists( 'wp_generate_uuid4' ) ) {
-			return wp_generate_uuid4();
-		}
-
-		return bin2hex( random_bytes( 16 ) );
 	}
 }
